### 2023-11-30 #6102 @egriendeau
> Hello @egriendeau, we'll use this pre-review issue to sort out any initial problems with the paper, and find an editor and reviewers.
>
> It looks like the citations in the paper are not building properly, because you have backticks (``) around all the citation commands—please remove those. The citation commands should just appear normally in the text source (e.g., "ARTIQ [@Bourdeauducq:2016]"), though make sure there is a space before the `[@` or `@`.
>
> In addition, can you check the missing DOI warnings above, and add any that are missing?
>
> Also, I noticed in the submission notes that you mentioned publications related to the software. Do any of those describe the software package itself?

I will fix the the citation back tick error and check on the DOIs. None of the publications descibe this software package itself.
