+++ paper.md
@@ -53,13 +53,13 @@ The QICK-DAWG package consists of FPGA firmware (modified from QICK firmware) co
 
 # Related Work
 
-Open-source software packages for quantum control and data acquisition such as ARTIQ`[@Bourdeauducq:2016]`, QubiC `[@ Xu:2023]`, and QICK`[@Stefanazzi:2022]` have been continually developed over the past decade for a variety of quantum experiments. ARTIQ--Advanced Real-Time Infrastructure for Quantum Physics--began in 2013 at the NIST ION Storage Group `[@Bourdeauducq:2016 ]`. ARTIQ was created in response to the shortcomings in both in-house and existing commercial based quantum control systems. ARTIQ helps describe complex experiments and runs on a high-level Python based programming language and is executed on an FPGA. ARTIQ has since partnered with the hardware project Sinara to improve quality and reproducibility of the hardware being used `[@Kulik:2022]`. The ARTIQ/Sinara software-hardware combination is focused on the fast control of ion-based quantum processors and has been used in nearly one thousand physics experiments `[@Shammah:2023]`.  
+Open-source software packages for quantum control and data acquisition such as ARTIQ [@Bourdeauducq:2016], QubiC [@ Xu:2023], and QICK [@Stefanazzi:2022] have been continually developed over the past decade for a variety of quantum experiments. ARTIQ--Advanced Real-Time Infrastructure for Quantum Physics--began in 2013 at the NIST ION Storage Group [@Bourdeauducq:2016 ]. ARTIQ was created in response to the shortcomings in both in-house and existing commercial based quantum control systems. ARTIQ helps describe complex experiments and runs on a high-level Python based programming language and is executed on an FPGA. ARTIQ has since partnered with the hardware project Sinara to improve quality and reproducibility of the hardware being used [@Kulik:2022]. The ARTIQ/Sinara software-hardware combination is focused on the fast control of ion-based quantum processors and has been used in nearly one thousand physics experiments [@Shammah:2023].  
 
 Focusing on hardware, Xilinx’s ZCU216, Xilinx’s ZCU111 and Real Digital’s RFSoC4x2 are recently available RFSoC FPGAs with the ability to generate control pulses at high frequencies (6–10GHz) and to digitize signals from photodiodes and single photon detector modules at high sample rates. The precise high frequencies pulse generation, readout capability, compact size, and relatively low cost of these RFSoC FPGAs make them ideal candidates for defect control hardware. There are two open-source quantum hardware software packages currently supporting the use of RFSoC’s: Qubic and QICK. 
 
-QubiC, started in 2018 at the Advanced Quantum Testbed at the Lawrence Berkeley National Laboratory, is an FPGA based open-source control and acquisition software, and it is focused on superconducting qubits. Recently, QubiC adopted Xilinx's ZCU216 to take advantage of new RFSoC capabilities `[@Xu:2021,@Xu:2023]`. 
+QubiC, started in 2018 at the Advanced Quantum Testbed at the Lawrence Berkeley National Laboratory, is an FPGA based open-source control and acquisition software, and it is focused on superconducting qubits. Recently, QubiC adopted Xilinx's ZCU216 to take advantage of new RFSoC capabilities [@Xu:2021,@Xu:2023]. 
 
-QICK (Quantum Instrumentation Control Kit) is the second package to utilize RFSoC FPGA’s, developed in 2021 `[@Stefanazzi:2022]`. QICK has been applied to superconducting, spin, atomic, molecular, and optical qubit systems, and has reached 40 labs in the last two years `[@Shammah:2023]`. Most appealing for quantum sensing, QICK has complied firmware for Real Digital’s RFSoC4x2 which is the lowest cost option of commercial off the shelf FPGA boards. 
[…]
