+++ paper.bib
@@ -67,7 +67,8 @@ keywords = {engineering software, quantum gate calibration, FPGA, qubit control,
   number={9},
   pages={2707--2747},
   year={2019},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={https://doi.org/10.5281/zenodo.1478113}
 }
 
 @article{Casola:2018,
@@ -78,7 +79,8 @@ keywords = {engineering software, quantum gate calibration, FPGA, qubit control,
   number={1},
   pages={1--13},
   year={2018},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={https://doi.org/10.1038/natrevmats.2017.88}
 }
 @article{Henshaw:2023,
 doi = {10.1088/2633-4356/ace095},
@@ -102,7 +104,8 @@ abstract = {The nitrogen-vacancy (NV) color center in diamond has demonstrated g
   volume={120},
   number={17},
   year={2022},
-  publisher={AIP Publishing}
+  publisher={AIP Publishing},
+  doi={https://doi.org/10.1063/5.0083774}
 }
 @article{Wang:2022,
   title={Noninvasive measurements of spin transport properties of an antiferromagnetic insulator},
@@ -112,7 +115,8 @@ abstract = {The nitrogen-vacancy (NV) color center in diamond has demonstrated g
   number={1},
   pages={eabg8562},
   year={2022},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={https://doi.org/10.1126/sciadv.abg8562}
 }
 @ARTICLE{Xu:2021,
   author={Xu, Yilun and Huang, Gang and Balewski, Jan and Naik, Ravi and Morvan, Alexis and Mitchell, Bradley and Nowrouzi, Kasra and Santiago, David I. and Siddiqi, Irfan},
