### 2023-11-30 #6102 @egriendeau
I fixed the citation ticks and added the DOIs. It looks like the citations are generating correctly in the proof pdf now.
