+++ .github/workflows/python-app.yaml
@@ -1,0 +1,56 @@
+# This workflow will install Python dependencies, run tests and lint with a single version of Python
+# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
+
+name: Python application
+
+on:
+  push:
+    branches: [ "main", "development"]
+  pull_request:
+    branches: [ "main", "development" ]
+  # Allow running manually from Actions tab
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v3
+      with:
+        python-version: "3.10"
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install flake8 pytest
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+    - name: Lint with flake8
+      run: |
+        # stop the build if there are Python syntax errors or undefined names
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+         - uses: actions/checkout@v2
+    # - name: autopep8
+    #   uses: peter-evans/autopep8@v2
+    #   with:
+    #     args: --in-place --recursive --select=E101,E126,E128,E225,E231,E261,E303,E402,W391 .
+    # - name: Create Pull Request
+    #   uses: peter-evans/create-pull-request@v3
+    #   with:
+    #     commit-message: autopep8 action fixes
+    #     title: Fixes by autopep8 action
+    #     body: This is an auto-generated PR with fixes by autopep8.
+    #     labels: autopep8, automated pr
+    #     reviewers: i-am-mounce
+    #     branch: autopep8-patches
+    
+    # - name: Test with pytest
+    #   run: |
+    #     pytest
