+++ qickdawg/nvpulsing/development/lockinodmrtwosources.py
@@ -4,6 +4,7 @@ from ..util import ItemAttribute
 
 import numpy as np
 
+
 class LockinODMRTwoSources(NVAveragerProgram):
 
     def initialize(self):
@@ -17,33 +18,32 @@ class LockinODMRTwoSources(NVAveragerProgram):
 
         # Get registers for mw1
         self.declare_gen(ch=self.cfg.mw1_channel, nqz=self.cfg.mw1_nqz)        
-        
+
         # Setup pulse defaults microwave
         self.set_pulse_registers(ch=self.cfg.mw1_channel,
-                                     style='const',
-                                     freq=self.cfg.mw1_start_freg,
-                                     gain= self.cfg.mw1_gain,
-                                     length=self.cfg.readout_treg,
-                                     phase=0)
-    
+                                 style='const',
+                                 freq=self.cfg.mw1_start_freg,
+                                 gain=self.cfg.mw1_gain,
+                                 length=self.cfg.readout_treg,
+                                 phase=0)
+
         # Get registers for mw1
         self.declare_gen(ch=self.cfg.mw2_channel, nqz=self.cfg.mw1_nqz)        
-        
+
         # Setup pulse defaults microwave
         self.set_pulse_registers(ch=self.cfg.mw2_channel,
-                                     style='const',
-                                     freq=self.cfg.mw2_start_freg,
-                                     gain= self.cfg.mw2_gain,
-                                     length=self.cfg.readout_treg,
-                                     phase=self.cfg.mw2_phase_preg)
+                                 style='const',
+                                 freq=self.cfg.mw2_start_freg,
+                                 gain=self.cfg.mw2_gain,
+                                 length=self.cfg.readout_treg,
+                                 phase=self.cfg.mw2_phase_preg)
 
+        # Get frequency register and convert frequency values to integers
+        self.mw1_frequency_register = self.get_gen_reg(self.cfg.mw1_channel, "freq")   
+        self.mw2_frequency_register = self.get_gen_reg(self.cfg.mw2_channel, "freq")   
 
-        ## Get frequency register and convert frequency values to integers
-        self.mw1_frequency_register=self.get_gen_reg(self.cfg.mw1_channel, "freq")   
-        self.mw2_frequency_register=self.get_gen_reg(self.cfg.mw2_channel, "freq")   
+        df_points = (self.cfg.mw1_end_freg - self.cfg.mw1_start_freg) // self.cfg.mw1_delta_freg + 1
 
-        df_points = (self.cfg.mw1_end_freg - self.cfg.mw1_start_freg)//self.cfg.mw1_delta_freg +1
-        
         self.add_sweep(NVQickSweep(self, self.mw1_frequency_register, self.cfg.mw1_start_freqMHz, self.cfg.mw1_end_freqMHz, df_points, source2=self.mw2_frequency_register))
 
         if self.cfg.pre_init:
@@ -54,23 +54,23 @@ class LockinODMRTwoSources(NVAveragerProgram):
         self.synci(400)  # give processor some time to configure pulses
 
     def body(self):
-            t0 = 0
+        t0 = 0
 
-            self.pulse(ch=self.cfg.mw1_channel, t=t0)
-            self.pulse(ch=self.cfg.mw2_channel, t=t0)
+        self.pulse(ch=self.cfg.mw1_channel, t=t0)
+        self.pulse(ch=self.cfg.mw2_channel, t=t0)
 
-            self.trigger(adcs=[0],
-                         adc_trig_offset=0)
+        self.trigger(adcs=[0],
+                     adc_trig_offset=0)
 
-            self.wait_all()
[…]
