+++ qickdawg/nvpulsing/development/cpmgxy8ndelaysweep.py
@@ -4,6 +4,7 @@ from ..util import ItemAttribute
 
 import numpy as np
 
+
 class CPMGXY8nDelaySweep(NVAveragerProgram):
     def initialize(self):
         """
@@ -18,60 +19,57 @@ class CPMGXY8nDelaySweep(NVAveragerProgram):
         # Get registers for mw
 
         self.declare_gen(ch=self.cfg.mw_channel, 
-                            nqz=self.cfg.mw_nqz)        
-        
+                         nqz=self.cfg.mw_nqz)        
+
         # Setup pulse defaults microwave
         self.default_pulse_registers(ch=self.cfg.mw_channel,
-                                    style='const',
-                                    freq=self.cfg.mw_freg,
-                                    length=self.cfg.mw_pi2_treg,
-                                    gain=self.cfg.mw_gain)
-    
+                                     style='const',
+                                     freq=self.cfg.mw_freg,
+                                     length=self.cfg.mw_pi2_treg,
+                                     gain=self.cfg.mw_gain)
+
         self.set_pulse_registers(ch=self.cfg.mw_channel,
                                  phase=0)
-        
-        
+
         # Delcare laser controls
         if self.cfg.laser_control == 'aom':
 
             self.declare_gen(ch=self.cfg.aom_channel, nqz=1)
 
             self.set_pulse_registers(ch=self.cfg.aom_channel, 
-                                        style='const',
-                                        freq=self.cfg.aom_freg,
-                                        gain=self.cfg.aom_gain,
-                                        length=self.cfg.laser_on_treg,
-                                        phase=0)
+                                     style='const',
+                                     freq=self.cfg.aom_freg,
+                                     gain=self.cfg.aom_gain,
+                                     length=self.cfg.laser_on_treg,
+                                     phase=0)
         elif self.cfg.laser_control == 'ttl':
             pass
         else:
             assert 0, "cfg.laser_control must be assigned to 'aom' or 'ttl'"
-
 
         # Addd loops
         self.delay_register = self.new_gen_reg(self.cfg.mw_channel,
                                                name='delay', 
                                                init_val=self.cfg.delay_start_treg)
         self.time_register = self.new_gen_reg(self.cfg.mw_channel,
-                                               name='time', 
-                                               init_val=0)
+                                              name='time', 
+                                              init_val=0)
 
-
-        if self.cfg.scaling_mode=='exponential':
+        if self.cfg.scaling_mode == 'exponential':
             self.add_sweep(NVQickSweep(self, 
-                                    self.delay_register,
-                                    self.cfg.delay_start_treg, 
-                                    self.cfg.delay_end_treg,
-                                    expts=self.cfg.nsweep_points,
-                                    scaling_mode= self.cfg.scaling_mode,
-                                    scaling_factor=self.cfg.scaling_factor))
+                                       self.delay_register,
+                                       self.cfg.delay_start_treg, 
+                                       self.cfg.delay_end_treg,
+                                       expts=self.cfg.nsweep_points,
+                                       scaling_mode=self.cfg.scaling_mode,
+                                       scaling_factor=self.cfg.scaling_factor))
[…]
