+++ qickdawg/nvpulsing/development/lockinodmrtwosources2.py
@@ -6,6 +6,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.image as mpimg
 
+
 class LockinODMRTwoSources2(NVAveragerProgram):
 
     def initialize(self):
@@ -22,11 +23,11 @@ class LockinODMRTwoSources2(NVAveragerProgram):
             self.declare_gen(ch=self.cfg.aom_channel, nqz=1)
 
             self.set_pulse_registers(ch=self.cfg.aom_channel, 
-                                        style='const',
-                                        freq=self.cfg.aom_freg,
-                                        gain=self.cfg.aom_gain,
-                                        length=self.cfg.readout_integration_treg,
-                                        phase=0)
+                                     style='const',
+                                     freq=self.cfg.aom_freg,
+                                     gain=self.cfg.aom_gain,
+                                     length=self.cfg.readout_integration_treg,
+                                     phase=0)
         elif self.cfg.laser_control == 'ttl':
             pass
         else:
@@ -34,35 +35,34 @@ class LockinODMRTwoSources2(NVAveragerProgram):
 
         # Get registers for mw 1
         self.declare_gen(ch=self.cfg.mw_channel, nqz=self.cfg.mw_nqz)        
-        
+
         # Setup pulse defaults microwave
         self.default_pulse_registers(ch=self.cfg.mw_channel,
                                      style='const',
-                                     gain= self.cfg.mw_gain,
+                                     gain=self.cfg.mw_gain,
                                      length=self.cfg.readout_integration_treg,
                                      phase=0)
-        
+
         self.set_pulse_registers(ch=self.cfg.mw_channel, 
                                  freq=self.cfg.mw_start_freg
-                                )
-
+                                 )
 
         # Get registers for mw2
         self.declare_gen(ch=self.cfg.mw2_channel, nqz=self.cfg.mw2_nqz)        
-        
+
         # Setup pulse defaults microwave
         self.default_pulse_registers(ch=self.cfg.mw2_channel,
                                      style='const',
-                                     gain= self.cfg.mw2_gain,
+                                     gain=self.cfg.mw2_gain,
                                      length=self.cfg.readout_integration_treg,
                                      phase=self.cfg.mw2_phase_preg)
 
         self.set_pulse_registers(ch=self.cfg.mw2_channel, 
                                  freq=self.cfg.mw_start_freg
-                                )
+                                 )
 
-        self.mw1_frequency_register=self.get_gen_reg(self.cfg.mw_channel, "freq")   
-        self.mw2_frequency_register=self.get_gen_reg(self.cfg.mw2_channel, "freq")   
+        self.mw1_frequency_register = self.get_gen_reg(self.cfg.mw_channel, "freq")   
+        self.mw2_frequency_register = self.get_gen_reg(self.cfg.mw2_channel, "freq")   
 
         self.add_sweep(NVQickSweep(self, 
                                    self.mw1_frequency_register,
@@ -70,7 +70,6 @@ class LockinODMRTwoSources2(NVAveragerProgram):
                                    self.cfg.mw_end_fMHz, 
                                    self.cfg.nsweep_points, 
                                    source2_reg=self.mw2_frequency_register))
-
 
         self.synci(400)  # give processor some time to self.cfgure pulses
 
@@ -80,22 +79,22 @@ class LockinODMRTwoSources2(NVAveragerProgram):
[…]

[…]
