+++ qickdawg/nvpulsing/hahnechodelaysweep.py
@@ -275,8 +275,11 @@ class HahnEchoDelaySweep(NVAveragerProgram):
                 str(cfg.readout_integration_tus)[:4]), fontsize=10)
             plt.text(850, 357, "relax_delay \n = {} us".format(str(cfg.relax_delay_tus)[:4]), fontsize=10)
             plt.text(400, 430, "laser_on = {} us".format(cfg.laser_on_tus), fontsize=12)
-            plt.text(375, 605, "    Sweep delay from {} us to {} us \n\
-                                     in {} {} steps".format(int(cfg.delay_start_tns), int(cfg.delay_end_tns), cfg.nsweep_points, cfg.scaling_mode), fontsize=12)
+            plt.text(
+                375, 605,
+                "    Sweep delay from {} us to {} us \n                                     in {} {} steps".format(
+                    int(cfg.delay_start_tns), int(cfg.delay_end_tns), 
+                    cfg.nsweep_points, cfg.scaling_mode), fontsize=12)
             plt.title("               Hahn Echo Pulse Sequence", fontsize=20)
 
     def time_per_rep(self):
