+++ .github/workflows/python-app.yaml
@@ -5,9 +5,9 @@ name: Python application
 
 on:
   push:
-    branches: [ "main", "development"]
+    branches: [ "main", "dev"]
   pull_request:
-    branches: [ "main", "development" ]
+    branches: [ "main", "dev" ]
   # Allow running manually from Actions tab
   workflow_dispatch:
 
@@ -36,7 +36,6 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-         - uses: actions/checkout@v2
     # - name: Test with pytest
     #   run: |
     #     pytest
