+++ .github/workflows/python-app.yaml
@@ -25,6 +25,13 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: "3.10"
+    - name: Install package
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install wheel
+        python -m pip install flake8
+        python -m pip install -e .
+        # python setup.py build_ext --inplace
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
