+++ paper.md
@@ -29,7 +29,7 @@ authors:
 - name: M. P. Lilly
   affiliation: 2 
 - name: A. M. Mounce
-  corresponding: true
+  corresponding: true amounce@sandia.gov
   orcid: 0000-0002-8115-2764
   affiliation: 2 
 affiliations:
