### 2024-01-09 #6102 @phibeck
Hi @egriendeau, thanks for your submission and for the reviewer suggestions. I'll be looking for reviewers next. As a side note, with six pages your paper is on the longer side (we aim for <= 1000 words, but not strict), so you could think about moving parts, e.g. the example features, into the repository directly.
