### 2024-01-09 #6102 @phibeck
@egriendeau Could you also please fix the DOIs (remove the extra 'https://doi.org/' prefix) Thanks!
