+++ paper.bib
@@ -31,7 +31,6 @@ publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
 volume = {4},
 number = {1},
-url = {https://doi.org/10.1145/3529397},
 doi = {10.1145/3529397},
 abstract = {As the size and complexity of a quantum computer increases, quantum bit (qubit) characterization and gate optimization become complex and time-consuming tasks. Current calibration techniques require complicated and verbose measurements to tune up qubits and gates, which cannot easily expand to the large-scale quantum systems. We develop a concise and automatic calibration protocol to characterize qubits and optimize gates using QubiC, which is an open source FPGA (field-programmable gate array)-based control and measurement system for superconducting quantum information processors. We propose multi-dimensional loss-based optimization of single-qubit gates and full XY-plane measurement method for the two-qubit CNOT gate calibration. We demonstrate the QubiC automatic calibration protocols are capable of delivering high-fidelity gates on the state-of-the-art transmon-type processor operating at the Advanced Quantum Testbed at Lawrence Berkeley National Laboratory. The single-qubit and two-qubit Clifford gate infidelities measured by randomized benchmarking are of 4.9(1.1) \texttimes{} 10-4 and 1.4(3) \texttimes{} 10-2, respectively.},
 journal = {ACM Transactions on Quantum Computing},
@@ -68,7 +67,7 @@ keywords = {engineering software, quantum gate calibration, FPGA, qubit control,
   pages={2707--2747},
   year={2019},
   publisher={Nature Publishing Group UK London},
-  doi={https://doi.org/10.5281/zenodo.1478113}
+  doi={10.5281/zenodo.1478113}
 }
 
 @article{Casola:2018,
@@ -80,7 +79,7 @@ keywords = {engineering software, quantum gate calibration, FPGA, qubit control,
   pages={1--13},
   year={2018},
   publisher={Nature Publishing Group},
-  doi={https://doi.org/10.1038/natrevmats.2017.88}
+  doi={10.1038/natrevmats.2017.88}
 }
 @article{Henshaw:2023,
 doi = {10.1088/2633-4356/ace095},
@@ -105,7 +104,7 @@ abstract = {The nitrogen-vacancy (NV) color center in diamond has demonstrated g
   number={17},
   year={2022},
   publisher={AIP Publishing},
-  doi={https://doi.org/10.1063/5.0083774}
+  doi={10.1063/5.0083774}
 }
 @article{Wang:2022,
   title={Noninvasive measurements of spin transport properties of an antiferromagnetic insulator},
@@ -116,7 +115,7 @@ abstract = {The nitrogen-vacancy (NV) color center in diamond has demonstrated g
   pages={eabg8562},
   year={2022},
   publisher={American Association for the Advancement of Science},
-  doi={https://doi.org/10.1126/sciadv.abg8562}
+  doi={10.1126/sciadv.abg8562}
 }
 @ARTICLE{Xu:2021,
   author={Xu, Yilun and Huang, Gang and Balewski, Jan and Naik, Ravi and Morvan, Alexis and Mitchell, Bradley and Nowrouzi, Kasra and Santiago, David I. and Siddiqi, Irfan},
