+++ docs/source/programs/readout_window.rst
@@ -1,0 +1,4 @@
+Readout Window
+==============
+
+.. autoclass:: qickdawg.ReadoutWindow
