### 2024-01-24 #6102 @egriendeau
@phibeck Can we suggest reviewers in the subject area but who have not reviewed a JOSS paper before? Also, I'm working on cutting the paper length down a bit. Thanks!
