+++ pyproject.toml
@@ -1,0 +1,5 @@
+[project]
+name = "qick-dawg"
+
+[tool.semantic_release]
+version_variables = ["setup.py:__version__"]
