+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "qick-dawg"
-version = "0.0.0"
+version = "0.0.1"
 dynamic = ["version"]
 
 
