### 2024-01-28 c3c0fd3
Configured docs for read the docs upload
