+++ docs/html/_sources/index.rst.txt
@@ -1,34 +1,0 @@
-.. qick-dawg documentation master file, created by
-   sphinx-quickstart on Tue Jan  9 15:25:09 2024.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-QICK-DAWG
-=========
-
-.. toctree::
-   :maxdepth: 2
-   :caption: The basics
-
-   thebasics/introduction
-   thebasics/getting_started
-
-.. toctree:: 
-   :maxdepth: 2
-   :caption: Programs/Pulse Sequences
-
-   programs/odmr
-   programs/readout_window
-   programs/rabi_oscillations
-   programs/ramsay_interferometry
-   programs/hahn_echo
-   programs/t1_relaxation
-   programs/cpmgxy8
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
-
