+++ docs/html/_sources/programs/readout_window.rst.txt
@@ -1,4 +1,0 @@
-Readout Window
-==============
-
-.. autoclass:: qickdawg.ReadoutWindow
