+++ docs/html/_sources/programs/odmr.rst.txt
@@ -1,6 +1,0 @@
-ODMR
-====
-
-ODMR is optically detected nuclear magnetic resonance
-
-.. autoclass:: qickdawg.LockinODMR
