+++ docs/html/_sources/programs/cpmgxy8.rst.txt
@@ -1,3 +1,0 @@
-CPMGXY-8
-====
-
