+++ docs/html/_sources/programs/t1_relaxation.rst.txt
@@ -1,4 +1,0 @@
-T1 Relaxation
-=============
-
-.. autoclass:: qickdawg.T1DelaySweep
