+++ docs/html/_sources/programs/rabi_oscillations.rst.txt
@@ -1,4 +1,0 @@
-Rabi Oscillations
-=====
-
-.. autoclass:: qickdawg.RabiSweep
