+++ .github/workflows/semantic_release.yaml
@@ -3,7 +3,7 @@ name: Semantic Release
 on:
   push:
     branches:
-      - master
+      - main
   workflow_dispatch:
 
 jobs:
