+++ paper.md
@@ -49,53 +49,28 @@ bibliography: paper.bib
 ---
 
 # Summary	
-The Quantum Instrumentation Control Kit - Defect Arbitrary Waveform Generator (QICK-DAWG), is an open-source software and firmware package for full quantum control and measurement of defect centers in diamond and other semiconductor materials for quantum sensing. QICK-DAWG extends the capabilities of the Quantum Instrumentation Control Kit (QICK, an open-source qubit firmware and software package) to defect centers by implementing controlled laser pulsing and low frequency readout required for defect initialization, control, and measurement using recently available Radio Frequency System-on-Chip (RFSoC) Field Programmable Gate Arrays (FPGA). In addition to user-friendly software and firmware, QICK-DAWG adds documentation that walks users through modifying the hardware of commercial off-the-shelf evaluation boards (from Real Digital and Xilinx) to digitize low frequency signals. QICK-DAWG’s measurement programs consist of specific microwave, laser, and readout pulse sequences built in a Python framework for consistency and extensibility. Pulse sequence programs and data analysis scripts are included to collect and characterize photoluminescence (PL) intensity, optically detected magnetic resonance (ODMR) spectra, PL readout windows, Rabi oscillations, Ramsay interference spectra, Hahn echo spin-spin relaxation times T<sub>2</sub>, and spin-lattice relaxation times T<sub>1</sub>. Additional pulse sequence programs and data analysis scripts will be added in the future. QICK-DAWG also implements live-update of plots for PL intensities to optimize laser alignment, broadband ODMR spectra for magnetic field alignment, and Rabi oscillations to optimize microwave antenna alignment or positioning. 
+The Quantum Instrumentation Control Kit - Defect Arbitrary Waveform Generator (QICK-DAWG), is an open-source software and firmware package for full quantum control and measurement of nitrogen-vacancy (NV) color-centers in diamond and other quantum defects in semiconductor materials for quantum sensing. QICK-DAWG extends the capabilities of the Quantum Instrumentation Control Kit (QICK, an open-source qubit firmware and software package) to quantum defects by implementing controlled laser pulsing and low frequency readout required for defect initialization, control, and measurement using recently available Radio Frequency System-on-Chip (RFSoC) Field Programmable Gate Arrays (FPGA). In addition to user-friendly software and firmware, QICK-DAWG adds documentation that guides users through hardware setup.  
 
-The QICK-DAWG package consists of FPGA firmware (modified from QICK firmware) compiled in Vivado, Python software that extends QICK for specific pulse programs and altered functionality, instructions for installation and hardware modifications, and a demo Jupyter notebook. The Readme.md file walks users through rudimentary hardware setup, installation, and modification required for low frequency data collection using RealDigital’s RFSoC4x2. The package also has a batch file with directions for easy installation setup for the RFSoC Linux kernel that drives firmware, controls, and reads from the FPGA.  The demo Jupyter Notebook walks users through typical experimental flow and the configuration of each pulse sequence. Each measurement program includes a method that checks whether all required configurations for pulsing programs are present and a method which provides a visual representation of the plot sequence. Ultimately, QICK-DAWG is an extensible firmware and software package for quantum sensing using NVs and other quantum defects using one control hub for consistency among different experimental setups or laboratories.  
[…]
