+++ qickdawg/nvpulsing/ramsey.py
@@ -12,7 +12,7 @@ from .nvqicksweep import NVQickSweep
 
 import matplotlib.pyplot as plt
 import matplotlib.image as mpimg
-
+import os 
 
 class Ramsey(NVAveragerProgram):
     '''
@@ -203,6 +203,7 @@ class Ramsey(NVAveragerProgram):
         pass
 
     def plot_sequence(cfg=None):
+        
         '''
         Function that plots the pulse sequence generated by this program
 
@@ -212,11 +213,14 @@ class Ramsey(NVAveragerProgram):
             If None, this plots the squence with configuration labels
             If a `.NVConfiguration` object is supplied, the configuraiton value are added to the plot
         '''
+        graphics_folder = os.path.join(os.path.dirname(__file__), '../../graphics')
+        image_path = os.path.join(graphics_folder, 'Ramsey.png')
+
 
         if cfg is None:
             plt.figure(figsize=(12, 12))
             plt.axis('off')
-            plt.imshow(mpimg.imread('../graphics/Ramsey.png'))
+            plt.imshow(mpimg.imread(image_path))
             plt.text(500, 700, "config.reps", fontsize=14)
             plt.text(270, 365, "delay", fontsize=10)
             plt.text(400, 385, "  config.readout_reference_start", fontsize=10)
@@ -233,7 +237,7 @@ class Ramsey(NVAveragerProgram):
         else:
             plt.figure(figsize=(12, 12))
             plt.axis('off')
-            plt.imshow(mpimg.imread('../graphics/Ramsey.png'))
+            plt.imshow(mpimg.imread(image_path))
             plt.text(450, 700, "Repeat {} times".format(cfg.reps), fontsize=14)
             plt.text(270, 365, "delay", fontsize=10)
             plt.text(385, 385, "  readout_reference_start = {} us".format(cfg.readout_reference_start_tus), fontsize=10)
