### 2024-02-13 #6102 @phibeck
Okay, thank you @jayich!

:wave: @14shreyasp would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
