### 2024-02-13 #6102 @phibeck
Hi @egriendeau a few more notes on the manuscript: In line 45 in the manuscript one reference isn't picked up. Could you please also check the formatting of the references in lines 134f. in the text, it looks like perhaps this isn't the intentional formatting. There are also still a couple of `https://doi.org/` too many in some of the references in the manuscript.

In the meantime, I'm still trying to find reviewers. If you have more suggestions, let me know.
