### 2024-02-14 #6102 @14shreyasp
> Okay, thank you @jayich!
>
> 👋 @14shreyasp would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Yes, I am happy to review it!
