+++ paper.md
@@ -55,7 +55,7 @@ Specifically, the QICK-DAWG   package consists of FPGA firmware (modified from t
 
 # Related Work
 
-Recent open-source software and commercially available RFSoC FPGA   evaluation boards provide a strong foundation for developing an open-source control measurement, software, and firmware package for NVs and other diamond quantum defects. Open-source software packages for quantum control and data acquisition such as ARTIQ [@Bourdeauducq:2016], Qubic  [@ Xu:2023], and QICK [@Stefanazzi:2022] have been continually developed over the past decade for a variety of quantum experiments. These packages have been developed in response to the shortcoming in both in-house and existing commercial based quantum control systems. Focusing on hardware, recently available RFSoC FPGAs including Xilinx’s ZCU216, Xilinx’s ZCU111 and Real Digital’s RFSoC4x2 can generate control pulses at high frequencies (6–10GHz) and digitize signals from photodiodes and single photon detector modules at high sample rates. The precise high-frequency pulse generation, readout capability, compact size, and relatively low cost of these RFSoC FPGAs make them ideal candidates for defect control hardware.
+Recent open-source software and commercially available RFSoC FPGA   evaluation boards provide a strong foundation for developing an open-source control measurement, software, and firmware package for NVs and other diamond quantum defects. Open-source software packages for quantum control and data acquisition such as ARTIQ [@Bourdeauducq:2016], Qubic  [@Xu:2023], and QICK [@Stefanazzi:2022] have been continually developed over the past decade for a variety of quantum experiments. These packages have been developed in response to the shortcoming in both in-house and existing commercial based quantum control systems. Focusing on hardware, recently available RFSoC FPGAs including Xilinx’s ZCU216, Xilinx’s ZCU111 and Real Digital’s RFSoC4x2 can generate control pulses at high frequencies (6–10GHz) and digitize signals from photodiodes and single photon detector modules at high sample rates. The precise high-frequency pulse generation, readout capability, compact size, and relatively low cost of these RFSoC FPGAs make them ideal candidates for defect control hardware.
 
 QICK provides firmware, a high-level Python user interface, and accessible inexpensive hardware making it an ideal platform for extension. Both QICK and Qubic     utilize RFSoC FPGAs, however only QICK provides firmware for Real Digital’s RFSoC4x2, the lowest cost commercial off the shelf FPGA board. QICK also provides a high-level Python user interface that supports rapid implementation and simple in lab parameter modification  .   Additionally, QICK is already a popular software package; QICK has been applied to superconducting, spin, atomic, molecular, and optical qubit systems, and has reached 40 labs in the last two years [@Shammah:2023]. Simple modification to QICK firmware, hardware and software is necessary for implementation of QICK in spin-based quantum control experiments like NVs and other diamond quantum defects. QICK-DAWG implements these modifications, extending QICK.  
 
