+++ paper.bib
@@ -9,9 +9,7 @@
     year = {2022},
     month = {04},
     issn = {0034-6748},
-    doi = {10.1063/5.0076249},
-    url = {https://doi.org/10.1063/5.0076249},
-    eprint = {https://pubs.aip.org/aip/rsi/article-pdf/doi/10.1063/5.0076249/16607898/044709\_1\_online.pdf},
+    doi = {10.1063/5.0076249}
 }
 
 @misc{Bourdeauducq:2016,
@@ -54,7 +52,8 @@ keywords = {engineering software, quantum gate calibration, FPGA, qubit control,
   title={Open Hardware in Quantum Technology},
   author={Shammah, Nathan and Roy, Anurag Saha and Almudever, Carmen G and Bourdeauducq, S{\'e}bastien and Butko, Anastasiia and Cancelo, Gustavo and Clark, Susan M and Heinsoo, Johannes and Henriet, Lo{\"\i}c and Huang, Gang and others},
   journal={arXiv preprint arXiv:2309.17233},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2309.17233}
 }
 
 
@@ -83,14 +82,13 @@ keywords = {engineering software, quantum gate calibration, FPGA, qubit control,
 }
 @article{Henshaw:2023,
 doi = {10.1088/2633-4356/ace095},
-url = {https://dx.doi.org/10.1088/2633-4356/ace095},
 year = {2023},
 month = {jul},
 publisher = {IOP Publishing},
 volume = {3},
 number = {3},
 pages = {035001},
-author = {Jacob Henshaw and Pauli Kehayias and Luca Basso and Michael Jaris and Rong Cong and Michael Titze and Tzu-Ming Lu and Michael P Lilly and Andrew M Mounce},
+author = {Jacob Henshaw and Pauli Kehayias and Luca Basso and Michael Jaris and Rong Cong and Michael Titze and {Tzu-Ming} Lu and Michael P Lilly and Andrew M Mounce},
 title = {Mitigation of nitrogen vacancy photoluminescence quenching from material integration for quantum sensing},
 journal = {Materials for Quantum Technology },
 abstract = {The nitrogen-vacancy (NV) color center in diamond has demonstrated great promise in a wide range of quantum sensing. Recently, there have been a series of proposals and experiments using NV centers to detect spin noise of quantum materials near the diamond surface. This is a rich complex area of study with novel nano-magnetism and electronic behavior, that the NV center would be ideal for sensing. However, due to the electronic properties of the NV itself and its host material, getting high quality NV centers within nanometers of such systems is challenging. Band bending caused by space charges formed at the metal-semiconductor interface force the NV center into its insensitive charge states. Here, we investigate optimizing this interface by depositing thin metal films and thin insulating layers on a series of NV ensembles at different depths to characterize the impact of metal films on different ensemble depths. We find an improvement of coherence and dephasing times we attribute to ionization of other paramagnetic defects. The insulating layer of alumina between the metal and diamond provide improved photoluminescence and higher sensitivity in all modes of sensing as compared to direct contact with the metal, providing as much as a factor of 2 increase in sensitivity, decrease of integration time by a factor of 4, for NV T 1 relaxometry measurements.}
