### 2024-02-19 #6380 @editorialbot
[REVIEW]: Quantum Instrumentation Control Kit - Defect Arbitrary Waveform Generator (QICK-DAWG): A Quantum Sensing Control Framework for Quantum Defects 
**Submitting author:** @egriendeau (<a href="http://orcid.org/0000-0002-0460-873X">Emmeline Riendeau</a>)
**Repository:** https://github.com/sandialabs/qick-dawg
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.1.0
