### 2024-02-21 49f436d
fixing ramsay to ramsey and adding hardware readme V1
