+++ docs/source/hardware/Readme.md
@@ -1,0 +1,325 @@
+# Quantum Instrumentation Control Kit - <br> Defect Arbitrary Waveform Generator 
+<p align="center">
+    <img src="./../_static/QD_logo.png"
+        alt="QD logo"
+        width="300x"/>
+</p>
+
+# Hardware Setup
+In this document we outline the hardware setup for an all-in-one RFSoC4x2 enclosure. We show:
+
+1. Enclosure Overview 
+
+    a. Components list 
+
+2. Enclosure Modification 
+
+    a. Drill Holes
+
+    b. 3D Printed Cages 
+
+    c. Low-Frequency Differential Amplifier 
+
+    d. BNC to PMOD TTL 
+
+3. Assembly 
+
+    a. Power 
+
+    b. Securing the RFSoC4x2
+
+    c. Securing 3D printed cages 
+
+    d. Front Panel Assembly
+
+    e. Input Signal Assembly 
+
+    f. Router and Ethernet Connection
+
+
+
+## 1. Enclosure Overview
+
+<p align="center">
+    <img src="./../_static/Enclosure_Schematic.png"
+        alt ="Fully assembled RFSoC4x2 Enclosure"
+        width="1000x"/>
+    
+</p>
+
+***1a. Components***
+- [Bud Industries CH-14404 Enclosure](https://www.digikey.com/en/products/detail/bud-industries/CH-14404/428959)
+- [RFSoC4x2 FPGA](https://www.xilinx.com/support/university/xup-boards/RFSoC4x2.html) 
+- [Router](https://www.amazon.com/TP-Link-Integrated-Lightening-Protection-TL-R605/dp/B08QTXNWZ1/ref=asc_df_B08QTXNWZ1/?tag=hyprod-20&linkCode=df0&hvadid=475692076734&hvpos=&hvnetw=g&hvrand=3761702075041011209&hvpone=&hvptwo=&hvqmt=&hvdev=c&hvdvcmdl=&hvlocint=&hvlocphy=1022494&hvtargid=pla-1149738264234&psc=1)
+- [DC Voltage Supply](https://www.digikey.com/en/products/detail/cui-inc/PYB30-Q24-T315-U/4477530) 
+- Low-Frequency Differential Amplifier [Texas Instruments LMH5401EVM](https://www.digikey.com/en/products/detail/texas-instruments/LMH5401EVM/5031896?s=N4IgTCBcDaIDIFkASBWALABgIwFEBqCIAugL5A)
+- 3D printed Router Cage
+- 3D printed Low-Frequency Differential Amplifier Holder
+- [Ethernet Pass-Through](https://www.digikey.com/en/products/detail/amphenol-ltw/RCP-5SPFFH-TCU7001/5253168)
+- [Power Pass-Through](https://www.digikey.com/en/products/detail/tensility-international-corp/54-00151/9829987)
+- [6 SMA Pass-Through](https://www.digikey.com/en/products/detail/amphenol-rf/132170/1011921) 
+- [8 BNC Pass-Through](https://www.digikey.com/en/products/detail/adam-tech/RF1-106-D-00-50-HDW/9830449)
+- [Pig Tail Power Cord](https://www.digikey.com/en/products/detail/tensility-international-corp/10-02937/9686418)
+- [1 x Flexible SMA Cable](https://www.digikey.com/en/products/detail/cinch-connectivity-solutions-johnson/415-0025-012/457078) 
+- [2 x Rigid SMA Cable](https://www.digikey.com/en/products/detail/cinch-connectivity-solutions-johnson/415-0081-006/1014585) 
[…]
