+++ docs/source/index.rst
@@ -20,10 +20,16 @@ QICK-DAWG
    programs/odmr
    programs/readout_window
    programs/rabi_oscillations
-   programs/ramsay_interferometry
+   programs/ramsey_interferometry
    programs/hahn_echo
    programs/t1_relaxation
    programs/cpmgxy8
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Hardware Setup
+   
+   hardware/Readme.md
 
 Indices and tables
 ==================
