+++ docs/source/programs/ramsey_interferometry.rst
@@ -1,0 +1,4 @@
+Ramsey Interferometry
+=====
+
+.. autoclass:: qickdawg.Ramsey
