+++ docs/source/conf.py
@@ -16,7 +16,7 @@ sys.path.insert(0, os.path.abspath('.'))
 
 # -- Project information -----------------------------------------------------
 
-project = 'qick-dawg'
+project = 'QICK-DAWG'
 copyright = '2024, Andy Mounce, Emmeline Riendeau'
 author = 'Andy Mounce'
 
@@ -51,12 +51,14 @@ autodoc_default_options = {'members': None}
 # a list of builtin themes.
 #
 html_theme = 'sphinx_rtd_theme'
+html_static_path = ['_static']
+html_logo = "_static/QD_logo-removebg-preview.png"
+html_theme_options = {
+    'logo-only': True,
+    'display_version': False,
+}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
 
-html_theme_options = {
-    'logo-only': False
-}
