+++ docs/source/programs/hahn_echo.rst
@@ -1,4 +1,8 @@
 Hahn Echo
-====
+=========
+
+.. image:: ./../_static/HAHN_ECHO.png
+    :width: 77%
+    :align: center
 
 .. autoclass:: qickdawg.HahnEchoDelaySweep
