+++ docs/source/programs/t1_relaxation.rst
@@ -1,4 +1,7 @@
 T1 Relaxation
 =============
+.. image:: ./../_static/T1.png
+    :width: 77%
+    :align: center
 
 .. autoclass:: qickdawg.T1DelaySweep
