+++ docs/source/programs/rabi_oscillations.rst
@@ -1,4 +1,7 @@
 Rabi Oscillations
-=====
+=================
+.. image:: ./../_static/RABI.png
+    :width: 77%
+    :align: center
 
 .. autoclass:: qickdawg.RabiSweep
