+++ docs/source/programs/ramsey_interferometry.rst
@@ -1,4 +1,8 @@
 Ramsey Interferometry
-=====
+=====================
+
+.. image:: ./../_static/Ramsey.png
+    :width: 77%
+    :align: center
 
 .. autoclass:: qickdawg.Ramsey
