+++ docs/source/thebasics/ReadMe.md
@@ -1,0 +1,110 @@
+# Introduction 
+
+## Quantum Instrumentation Control Kit - <br/>Defect Arbitrary Waveform Generator 
+<p align="center">
+    <img src="./../_static/QD_logo.png"
+        alt="QD logo"
+        width="300x"/>
+</p>
+
+
+## Description
+The Quantum Instrumentation Control Kit - Defect Arbitrary Waveform Generator (QICK-DAWG) is an extension of [QICK](https://github.com/openquantumhardware/qick) with the purpose of controlling nitrogen-vacancy in diamond and other quantum defects for quantum sensing and other quantum applications.  QICK-DAWG utilizes Xilinx Radio Frequency System-on-Chip (RFSoC) Field Programmable Gate Array (FPGAs) evaluation boards which are significantly less expensive than hardware currently available and have the potential for superior capabilities.
+
+In this package we supply specific mw, laser, and readout pulsing code to characterize  photoluminescence intensity, optically detected magnetic resonance (ODMR), spin readout, Rabi Oscillation, Hahn Echo T<sub>2</sub>, and T<sub>1</sub> relaxation. A Jupyter Notebook Demo is included for performing all of these measurements with an RFSoC4x2 to demonstrate typical measurements. 
+
+## Hardware
+
+- Either [RFSoC4x2](https://www.xilinx.com/support/university/xup-boards/RFSoC4x2.html)  or [ZCU111](https://www.xilinx.com/products/boards-and-kits/zcu111.html)
+- Router [(example)](https://www.amazon.com/TP-Link-Integrated-Lightening-Protection-TL-R605/dp/B08QTXNWZ1/ref=asc_df_B08QTXNWZ1/?tag=hyprod-20&linkCode=df0&hvadid=475692076734&hvpos=&hvnetw=g&hvrand=3761702075041011209&hvpone=&hvptwo=&hvqmt=&hvdev=c&hvdvcmdl=&hvlocint=&hvlocphy=1022494&hvtargid=pla-1149738264234&psc=1) and ethernet cords
+- Control Computer 
+- Low Frequency Differential Amplifier such as  [Texas Instruments LMH5401EVM](https://www.digikey.com/en/products/detail/texas-instruments/LMH5401EVM/5031896?s=N4IgTCBcDaIDIFkASBWALABgIwFEBqCIAugL5A)
+- 3 Voltage supplies to power the differential amplifier
+
+## Installation
+To install QICKDAWG software follow the README found [here](https://gitlab-ex.sandia.gov/amounce/qickdawg/-/blob/main/installation/RFSoC4x2_ReadMe.md#quantum-instrumentation-control-kit-defect-arbitary-waveform-generator). Software installation includes:
+
+## Usage
+QICKDAWG has programs for the characterization of:
+- Photoluminescence (PL) Intensity
+- Optically Detected Magnetic Resonance (ODMR)
+- Readout Window
+- Rabi Oscillations 
+- Hanhn Echo T<sub>2</sub>
+- T<sub>1</sub> Relaxation
+
+using a ZCU111 or RFSoC4x2. Support for the ZCU216 coming soon. 
+
+<!-- ### Example use: ###
+
+Optically Detected Magnetic Resonance (ODMR) is one of the QICKDAWG demo pulse sequences. To run ODMR, and any of the other pulse sequences, you must first set the default configuration. The following code, found in the demo Jupyter Notebook, sets the deault configuration.
+<p align="center">
+    <img src="./../_static/config_example.PNG"
+        alt="configuration code example"
+        width="500x"/>
+</p>
+
[…]
