+++ docs/requirements.txt
@@ -1,1 +1,2 @@
 sphinx-rtd-theme
+myst_parser
