### 2024-02-22 #6380 @ktahar
I have a few concerns before starting the review.
Substantial part of this project may be the firmware for specific hardware (RFSoC eval board) and I don't have it now.
Situation is similar to the case of pySLM2 #6205.

@phibeck Can I proceed even in this situation? Maybe we must ask the authors to create some mocks to emulate the hardware on PC, in order to check if the software part works as claimed.

@egriendeau Can you possibly include the "modified qick" part in the repository using git submodule (instead of just putting the qick directory)? I would like to see the difference between your modified qick and the original. Having a separate qick repository (with history of original qick) makes this much easier.
