### 2024-02-28 #6380 @egriendeau
Additionally, I corrected the link from the main readme to the installation readme
