### 2024-03-03 #6380 @ktahar
@ammounce Thank you for explanation. I will start my review by playing with the notebook. As I have minimal experience of
diamond NV experiments, I think I can get the feel for it.
