### 2024-03-04 #6380 @phibeck
Thank you, @egriendeau for fixing the link and @ammounce for clarifying the hardware requirements. Sounds like it should be doable to assess the software via the notebook to a good extent.

Thanks @sidihamady, @14shreyasp and @ktahar for getting the review started.  Let me know if you have any questions as you work through your checklists.
