### 2024-03-04 #6380 @ammounce
@ktahar @sidihamady, @14shreyasp A little more clarification: you **do** need the hardware to run the demo notebook too.

My comment was that perhaps the demo notebook (as previously) run, could be sufficient. However, if not, let us know and we'll try to setup a test server.  I believe this is possible, but could be challenging due to the fact that we're at at National Laboratory and internet security is tight.  Nonetheless, I might have a path to making a full test server.
