### 2024-03-05 #6380 @ktahar
Hi @ammounce , Thank you for clarification. I was misunderstanding that demo notebook can be run without hardware.
But I will try to start the review anyway.
