### 2024-03-05 fb1a0f1
pyproject.toml: don't exclude qickdawg package
- "qick*" excludes the qickdawg package which should be installed
