### 2024-03-05 sandialabs/qick-dawg#23 @ktahar
errors/warnings on docs building
I tried to build the docs in cloned repository.
When I hit `make html` in the `docs` directory, bunch of errors and warnings are printed (as pasted at the bottom).
Can these be fixed?

And is there a URL where I can browse the compiled html documentation?

```
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/hahnechodelaysweep.py:docstring of qickdawg.nvpulsing.hahnechodelaysweep.HahnEchoDelaySweep:8: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/hahnechodelaysweep.py:docstring of qickdawg.nvpulsing.hahnechodelaysweep.HahnEchoDelaySweep:10: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/hahnechodelaysweep.py:docstring of qickdawg.nvpulsing.hahnechodelaysweep.HahnEchoDelaySweep:11: WARNING: Block quote ends without a blank line; unexpected unindent.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/hahnechodelaysweep.py:docstring of qickdawg.nvpulsing.hahnechodelaysweep.HahnEchoDelaySweep:36: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/lockinodmr.py:docstring of qickdawg.nvpulsing.lockinodmr.LockinODMR:10: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/lockinodmr.py:docstring of qickdawg.nvpulsing.lockinodmr.LockinODMR.analyze_results:10: WARNING: Inline emphasis start-string without end-string.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/lockinodmr.py:docstring of qickdawg.nvpulsing.lockinodmr.LockinODMR.body:4: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/lockinodmr.py:docstring of qickdawg.nvpulsing.lockinodmr.LockinODMR.body:5: WARNING: Block quote ends without a blank line; unexpected unindent.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/rabisweep.py:docstring of qickdawg.nvpulsing.rabisweep.RabiSweep:11: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/rabisweep.py:docstring of qickdawg.nvpulsing.rabisweep.RabiSweep:32: WARNING: Definition list ends without a blank line; unexpected unindent.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/ramsey.py:docstring of qickdawg.nvpulsing.ramsey.Ramsey:8: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/ramsey.py:docstring of qickdawg.nvpulsing.ramsey.Ramsey:10: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/ramsey.py:docstring of qickdawg.nvpulsing.ramsey.Ramsey:11: WARNING: Block quote ends without a blank line; unexpected unindent.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/ramsey.py:docstring of qickdawg.nvpulsing.ramsey.Ramsey:36: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/readoutwindow.py:docstring of qickdawg.nvpulsing.readoutwindow.ReadoutWindow:12: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/readoutwindow.py:docstring of qickdawg.nvpulsing.readoutwindow.ReadoutWindow:33: WARNING: Definition list ends without a blank line; unexpected unindent.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/t1delaysweep.py:docstring of qickdawg.nvpulsing.t1delaysweep.T1DelaySweep:11: ERROR: Unexpected indentation.
/home/ktaha/dev/qick-dawg/qickdawg/nvpulsing/t1delaysweep.py:docstring of qickdawg.nvpulsing.t1delaysweep.T1DelaySweep:32: WARNING: Definition list ends without a blank line; unexpected unindent.
WARNING: unsupported theme option 'logo-only' given
/home/ktaha/dev/qick-dawg/docs/source/thebasics/ReadMe.md:77: WARNING: 'myst' cross-reference target not found: 'CINT' [myst.xref_missing]
/home/ktaha/dev/qick-dawg/docs/source/thebasics/ReadMe.md:77: WARNING: 'myst' cross-reference target not found: 'CINT' [myst.xref_missing]
/home/ktaha/dev/qick-dawg/docs/source/thebasics/ReadMe.md:77: WARNING: 'myst' cross-reference target not found: 'Hav' [myst.xref_missing]
```

https://github.com/openjournals/joss-reviews/issues/6380
