+++ docs/build/_sources/programs/t1_relaxation.rst.txt
@@ -1,0 +1,7 @@
+T1 Relaxation
+=============
+.. image:: ./../_static/T1.png
+    :width: 77%
+    :align: center
+
+.. autoclass:: qickdawg.T1DelaySweep
