+++ docs/build/_sources/programs/ramsey_interferometry.rst.txt
@@ -1,0 +1,8 @@
+Ramsey Interferometry
+=====================
+
+.. image:: ./../_static/Ramsey.png
+    :width: 77%
+    :align: center
+
+.. autoclass:: qickdawg.Ramsey
