+++ paper.md
@@ -1,87 +1,0 @@
----
-title: 'Quantum Instrumentation Control Kit – Defect Arbitrary Waveform Generator (QICK-DAWG): A Quantum Sensing Control Framework for Quantum Defects'
-tags:
-  - QICK
-  - Nitrogen-Vacancy Centers
-  - NVs 
-  - quantum defect characterization
-  - RFSoC4x2
-authors:
-- name: E. G. Riendeau
-  corresponding: emmelineriendeau@gmail.com
-  orcid: 0000-0002-0460-873X
-  affiliation: "1, 2" # (Multiple affiliations must be quoted)
-- name: L. Basso
-  affiliation: 2
-- name: J. J. Mah
-  affiliation: "2,3"
-- name: R. Cong
-  affiliation: "2,3" 
-- name: M. A. Sadi
-  affiliation: "2,4"
-- name: J. Henshaw
-  affiliation: 2 
-- name: K. M. Azizur-Rahman
-  affiliation: 2 
-- name: A. Jones 
-  affiliation: "2,5"
-- name: G. Joshi
-  affiliation: 2 
-- name: M. P. Lilly
-  affiliation: 2 
-- name: A. M. Mounce
-  corresponding: amounce@sandia.gov
-  orcid: 0000-0002-8115-2764
-  affiliation: 2 
-affiliations:
-- name: Haverford College, US
-  index: 1
-- name: Center for Integrated Nanotechnologies (CINT), Sandia National Laboratories, US
-  index: 2
-- name: Brown University, US
-  index: 3
-- name: Purdue University, US
-  index: 4
-- name: Georgia Institute of Technology, US
-  index: 5
-date: 27 November 2023
-bibliography: paper.bib       
----
-
-# Summary	
-The Quantum Instrumentation Control Kit - Defect Arbitrary Waveform Generator (QICK-DAWG), is an open-source software and firmware package for full quantum control and measurement of nitrogen-vacancy (NV) color-centers in diamond and other quantum defects in semiconductor materials for quantum sensing. QICK-DAWG extends the capabilities of the Quantum Instrumentation Control Kit (QICK, an open-source qubit firmware and software package) to quantum defects by implementing controlled laser pulsing and low frequency readout required for defect initialization, control, and measurement using recently available Radio Frequency System-on-Chip (RFSoC) Field Programmable Gate Arrays (FPGA). In addition to user-friendly software and firmware, QICK-DAWG adds documentation that guides users through hardware setup.  
-
[…]
