+++ paper.bib
@@ -1,126 +1,0 @@
-
-@article{ Stefanazzi:2022,
-    author = {Stefanazzi, Leandro and Treptow, Kenneth and Wilcer, Neal and Stoughton, Chris and Bradford, Collin and Uemura, Sho and Zorzetti, Silvia and Montella, Salvatore and Cancelo, Gustavo and Sussman, Sara and Houck, Andrew and Saxena, Shefali and Arnaldi, Horacio and Agrawal, Ankur and Zhang, Helin and Ding, Chunyang and Schuster, David I.},
-    title = "{The QICK (Quantum Instrumentation Control Kit): Readout and control for qubits and detectors}",
-    journal = {Review of Scientific Instruments},
-    volume = {93},
-    number = {4},
-    pages = {044709},
-    year = {2022},
-    month = {04},
-    issn = {0034-6748},
-    doi = {10.1063/5.0076249}
-}
-
-@misc{Bourdeauducq:2016,
-url={https://zenodo.org/records/51303},
-  title={ARTIQ 1.0 (1.0)},
-  author={Bourdeauducq, Sebastian and J{\"o}rdens, R and Zotov, P and Britton, J and Slichter, D and Leibrandt, D and Allcock, D and Hankin, A and Kermarrec, F and Sionneau, Y and others},
-  year={2016},
-  publisher={Zenodo}
-}
-
-@article{ Xu:2023,
-author = {Xu, Yilun and Huang, Gang and Balewski, Jan and Morvan, Alexis and Nowrouzi, Kasra and Santiago, David I. and Naik, Ravi K. and Mitchell, Brad and Siddiqi, Irfan},
-title = {Automatic Qubit Characterization and Gate Optimization with QubiC},
-year = {2022},
-issue_date = {March 2023},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-volume = {4},
-number = {1},
-doi = {10.1145/3529397},
-abstract = {As the size and complexity of a quantum computer increases, quantum bit (qubit) characterization and gate optimization become complex and time-consuming tasks. Current calibration techniques require complicated and verbose measurements to tune up qubits and gates, which cannot easily expand to the large-scale quantum systems. We develop a concise and automatic calibration protocol to characterize qubits and optimize gates using QubiC, which is an open source FPGA (field-programmable gate array)-based control and measurement system for superconducting quantum information processors. We propose multi-dimensional loss-based optimization of single-qubit gates and full XY-plane measurement method for the two-qubit CNOT gate calibration. We demonstrate the QubiC automatic calibration protocols are capable of delivering high-fidelity gates on the state-of-the-art transmon-type processor operating at the Advanced Quantum Testbed at Lawrence Berkeley National Laboratory. The single-qubit and two-qubit Clifford gate infidelities measured by randomized benchmarking are of 4.9(1.1) \texttimes{} 10-4 and 1.4(3) \texttimes{} 10-2, respectively.},
-journal = {ACM Transactions on Quantum Computing},
-month = {oct},
-articleno = {3},
-numpages = {12},
-keywords = {engineering software, quantum gate calibration, FPGA, qubit control, NISQ, gateware}
-}
-
-@INPROCEEDINGS{Kulik:2022,
[…]
