+++ docs/source/conf.py
@@ -54,7 +54,7 @@ html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
 html_logo = "_static/qickdawg_forgithub.png"
 html_theme_options = {
-    'logo-only': True,
+    'logo_only': True,
     'display_version': False,
 }
 
