+++ docs/source/programs/hahn_echo.rst
@@ -1,8 +1,10 @@
 Hahn Echo
 =========
-
+import os
+import sys
+sys.path.insert(0, os.path.abspath('../..'))
 .. image:: ./../_static/HAHN_ECHO.png
     :width: 77%
     :align: center
 
-.. autoclass:: qick-dawg.HahnEchoDelaySweep
+.. autoclass:: qickdawg.HahnEchoDelaySweep
