+++ docs/source/conf.py
@@ -31,8 +31,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.coverage',
     'sphinx_rtd_theme',
-    'myst_parser',
-    'numpy'
+    'myst_parser'
 ]
 
 # Add any paths that contain templates here, relative to this directory.
