+++ docs/source/programs/hahn_echo.rst
@@ -1,6 +1,6 @@
 Hahn Echo
 =========
-s
+
 .. image:: ./../_static/HAHN_ECHO.png
     :width: 77%
     :align: center
