+++ docs/source/programs/conf.py
@@ -65,5 +65,5 @@ html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
 
 html_theme_options = {
-    'logo_only': True
+    'logo-only': False
 }
