+++ docs/source/conf.py
@@ -63,7 +63,7 @@ html_theme = 'sphinx_rtd_theme'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
-# html_logo = "_static/qickdawg_forgithub.png"
-# html_theme_options = {
-#     'logo_only': True
-# }
+html_logo = "_static/qickdawg_forgithub.png"
+html_theme_options = {
+    'logo_only': True
+}
