+++ installation/Readme.md
@@ -1,5 +1,5 @@
 # RFSoC4x2 Setup 
-The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Digital](https://www.realdigital.org/) using AMD’s ZYNQ Ultrascale+ Gen 3 RFSoC ZU48DR chip.  While the ZU48DR has 8 digital-to-analog converters (DACs) and analog-to-digital converters (ADCs), the RFSOC4x2 only uses 4 DACs (5 GSa/s) and 2 ADCs (9.85 GSa/s). Nonetheless, this number of inputs and outputs is nearly perfect for NV and quantum defect control. However, as the RFSOC4x2 is sold, the ADCs have a high frequency 1GHz high-pass balun inline which is tyipcally too high frequency for our measurements and thus must be modified.
+The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Digital](https://www.realdigital.org/) using AMD’s ZYNQ Ultrascale+ Gen 3 RFSoC ZU48DR chip.  While the ZU48DR has eight digital-to-analog converters (DACs) and analog-to-digital converters (ADCs), the RFSOC4x2 only uses four DACs (5 GSa/s) and two ADCs (9.85 GSa/s). Nonetheless, this number of inputs and outputs is nearly perfect for NV and quantum defect control. However, as the RFSOC4x2 is sold, the ADCs have a high frequency 1GHz high-pass balun inline which is tyipcally too high frequency for our measurements and thus must be modified.
 
 
 <p align="center">
@@ -13,7 +13,7 @@ The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Di
 In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Specifically, we show how to:
 
 1. Setup RFSoC4x2 Hardware<br>
-    a. Bipass/Remove the balun & capacitors<br>
+    a. Bypass/Remove the balun & capacitors<br>
     b. Connect the lw frequency differential amplifier<br>
     c. Connect PMOD digial outputs<br>
     d. Assembling and powering on your RFSoC4x2 board
@@ -35,7 +35,7 @@ In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Speci
 - Low frequency differential amplifier [Texas Instruments LMH5401EVM](https://www.digikey.com/en/products/detail/texas-instruments/LMH5401EVM/5031896?s=N4IgTCBcDaIDIFkASBWALABgIwFEBqCIAugL5A)
 - 3 x DC output voltage supply (+3.2, +0.7, -1.8V for biasing the differential amplifier)
 - SMA cables
-- Computer with Ethernet port or ethernet adaptor
+- Computer with Ethernet port or Ethernet adaptor
 - Ethernet cord(s) (at least one from RFSoC to router)
 - Micro SD card reader
 ### Software 
@@ -45,9 +45,9 @@ In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Speci
     - [Pyro4](https://pypi.org/project/Pyro4/)
     - [Serpent](https://pypi.org/project/serpent/) 
 
-## 1a. Bipass/Remove the Balun & Capacitors
+## 1a. Bypass/Remove the Balun & Capacitors
 
-The signal measured by the photodiodes cannot be direcetly connected to the ADCs on the board, thus requiring modification the to the RFSoC4x2. The ADCs on the RFSoC4x2 have baluns and capacitors that act as high pass filters. When using photodiodes for photoluminesence detection, the signal is at low frequency thus the balun and capacitors on the board need to be removed or bipassed in order to get the signal into the ADC. Furthermore, as the ADCs take in a differential voltage signal, we have to add a differential amplifier which takes the signal from the photodetector in and outputs a biased signal to the ADCs for digitization (see section 1.b below). 
[…]
