### 2024-05-01 660fdc3
revising the installation readme typos
