+++ installation/Readme.md
@@ -1,4 +1,4 @@
-# RFSoC4x2 Setup 
+![PMOD](https://github.com/sandialabs/qick-dawg/assets/141430553/e0c483e8-3a90-4390-bf4f-2176c191e800)# RFSoC4x2 Setup 
 The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Digital](https://www.realdigital.org/) using AMD’s ZYNQ Ultrascale+ Gen 3 RFSoC ZU48DR chip.  While the ZU48DR has eight digital-to-analog converters (DACs) and analog-to-digital converters (ADCs), the RFSOC4x2 only uses four DACs (5 GSa/s) and two ADCs (9.85 GSa/s). Nonetheless, this number of inputs and outputs is nearly perfect for NV and quantum defect control. However, as the RFSOC4x2 is sold, the ADCs have a high frequency 1GHz high-pass balun inline which is typically too high frequency for our measurements and thus must be modified.
 
 
@@ -107,7 +107,7 @@ Note that you can instead leave the balun in place and directly solder coax cabl
 To control the laser through TTL you must connect your laser to the PMOD located on the corner of the board. To connect, we cut the female head off a PMOD cable and soldered on a female BNC head instead. PMOD A 1-8 are enabled for QICK-DAWG (in the demo we use PMOD 1). The image below provides a schematic of the PMOD on the RFSoC4x2.
 
 <p align="center">
-    <img src="graphics/PMOD.PNG"
+    <img src="graphics/PMOD.png"
         alt="PMOD diagram"
         width="800px"/>
 </p>
