### 2024-05-08 f8688c7
Delete qickdawg/nvpulsing/DEER.py
