### 2024-05-08 2f771d0
updating readme capacitor instructions
