### 2024-05-29 3b47fd6
adding cpmg xy 8 n
