+++ qickdawg/nvpulsing/CPMGXY8-n.py
@@ -1,0 +1,316 @@
+'''
+CPMGXY8
+=======================================================================
+An NVAveragerProgram class acquires data for CPMGXY8, 
+'''
+
+from .nvaverageprogram import NVAveragerProgram
+from .nvqicksweep import NVQickSweep
+
+import matplotlib.pyplot as plt
+import matplotlib.image as mpimg
+import os 
+
+class CPMGXY8(NVAveragerProgram):
+    '''
+    An NVAveragerProgram class that generates and executes a sequence used
+    to measure CPMGXY8
+
+    Parameters
+    ----------
+    soccfg : `qick.QickConfig`
+    cfg : `.NVConfiguration`
+        instance of `.NVConfiguration` class with attributes:
+        .adc_channel : int
+            ADC channel for gathering data, usually 0 or 1
+        .mw_channel : int
+            qick channel that provides microwave excitation
+            0 or 1 for RFSoC4x2
+            0 to 6 for ZCU111 or ZCU216
+        .mw_nqz : int
+            nyquist zone for microwave generator (1 or 2)
+        .mw_gain : int
+            gain of micrwave channel, in register values, from 0 to 2**15-1
+
+        .pre_init : bool
+            boolian value that indicates whether to pre-pulse the laser to initialize
+            the spin state
+
+        .relax_delay_treg : int
+            Time between on/off cycles and reps in register units
+        .readout_length_treg : int
+            Time for which the adc accumulates data in register units
+        .laser_readout_offset_treg : int
+            Time offset between initial laser triggering and readout start
+
+        .laser_gate_pmod : int
+            PMOD channel used to trigger laser source usually 0 to 6
+
+    Returns
+    `.CPMGXY8`
+        An instance of CMPGXY8 with compiled assembly language
+
+    Methods
+    -------
+    initialize
+        method that generates the assembly code that setups the adcs &  mw generators,
+        and performs other one-off setps
+    body
+        method that generates the assembly code that exectues in the middle of each sweep
+        and rep
+    plot_sequence
+        generates a plot labeled with self.cfg attributes or the required inputs
+    time_per_rep
+        returns the approximatetime for one rep to complete
+    total_time
+        returns the approximate total time for the entire program to complete
+    '''
+    required_cfg = [
+        "adc_channel",
+        "readout_integration_treg",
+        "mw_channel",
+        "mw_nqz",
+        "mw_pi2_treg",
+        "mw_gain",
+        "scaling_mode",
+        "delay_start_treg",
+        "delay_end_treg",
+        "nsweep_points",
+        "pre_init",
+        "laser_gate_pmod",
+        "laser_on_treg",
+        "relax_delay_treg",
+        "readout_delay_treg",
+        "reps",
+        "readout_reference_start_treg",
+        "laser_readout_offset_treg",
+        "mw_readout_delay_treg"]
+
+    def initialize(self):
+        '''Method that generates the assembly code that is sets up adcs and sources.
+
+        For HahnEchoDelaySweep this:
+        1. Configures the adc to acquire points for self.cfg.readout_integration_t#
+        2. Configures the microwave channel
+        3. Configures the sweep parameters
+        4. Initiailzes the spin state with a laser pulse
+        '''
+        self.check_cfg()
+
+        self.declare_readout(ch=0,
+                             freq=0,
+                             length=self.cfg.readout_integration_treg,
[…]
