### 2024-05-29 f4eef5f
writing cpmg xy n
