### 2024-06-06 7039aa9
fixing my email in toml and fixing the demo so it doesn't say pyscan
