+++ installation/Readme_Wifi.md
@@ -1,0 +1,247 @@
+# RFSoC4x2 setup (RFSoC4x2 Wifi)
+This Readme supports setting up your RFSoC4x2 where the RFSoC4x2 is connected directly to the internet. We have an alternative Readme for labs where the RFSoC4x2 is not permitted to be connected directly to the internet.s
+
+The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Digital](https://www.realdigital.org/) using AMD’s ZYNQ Ultrascale+ Gen 3 RFSoC ZU48DR chip.  While the ZU48DR has eight digital-to-analog converters (DACs) and analog-to-digital converters (ADCs), the RFSOC4x2 only uses four DACs (5 GSa/s) and two ADCs (9.85 GSa/s). Nonetheless, this number of inputs and outputs is nearly perfect for NV and quantum defect control. However, as the RFSOC4x2 is sold, the ADCs have a high frequency 1GHz high-pass balun inline which is typically too high frequency for our measurements and thus must be modified.
+
+
+<p align="center">
+    <img src="graphics/RFSoC4x2_base_image.png"
+        alt="RFSoC4x2 with inputs and outputs labelled"
+        width="600px"/>
+
+</p>
+
+
+In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Specifically, we show how to:
+
+1. Setup RFSoC4x2 Hardware<br>
+    a. Bypass/Remove the balun & capacitors<br>
+    b. Connect the low frequency differential amplifier<br>
+    c. Connect PMOD digital outputs<br>
+    d. Assembling and powering on your RFSoC4x2 
+    e. (Optional) Full enclosure
+
+2. Setup lab control computer
+    a. Download/clone QICK-DAWG
+    b. Install necessary packages
+
+3. Upload and install QICK-DAWG and other software to you RFSoC4x2<br>
+    a. Flashing your microSD card<br>
+    b. Cloning QICK-DAWG on your RFSoC4x2 
+    c. Installing necessary packages on your RFSOC4x2
+    d. Running the Pyro server to remotely connect to QICK and the RFSoC4x2
+    e. (Supplemental) Establishing SSH Remote Connection with the RFSoC4x2
+
+
+
+
+# 1. Setup RFSoC4x2 Hardware 
+## ***Prerequisites***
+
+- [RFSoC4x2](https://www.xilinx.com/support/university/xup-boards/RFSoC4x2.html) (with 12 volt 50 watt power supply) 
+- Low frequency differential amplifier [Texas Instruments LMH5401EVM](https://www.digikey.com/en/products/detail/texas-instruments/LMH5401EVM/5031896?s=N4IgTCBcDaIDIFkASBWALABgIwFEBqCIAugL5A)
+- 3 x DC output voltage supply (+3.2, +0.7, -1.8V for biasing the differential amplifier)
+- SMA cables
+- Ethernet cord(s) (at least one from RFSoC to router)
+- Micro SD card reader
+### Software 
+- [Win32DiskImager](https://sourceforge.net/projects/win32diskimager/) for Windows or Disk Manager on MacOS
+- Dependent packages (follow `Installing Necessary Packages` section--included in batch file setup)
+    - [QICK](https://github.com/openquantumhardware/qick)
+    - [Pyro4](https://pypi.org/project/Pyro4/)
+    - [Serpent](https://pypi.org/project/serpent/) 
+
+## 1a. Bypass/Remove the Balun & Capacitors
+
[…]
