### 2024-06-14 dbb41ff
updating WAN readme
