+++ installation/Readme.md
@@ -1,5 +1,5 @@
-# RFSoC4x2 setup (WAN - direct internet connection)
-This Readme supports setting up your RFSoC4x2 where the RFSoC4x2 is connected directly to the internet--wide area network (WAN). We have an alternative Readme for labs where the RFSoC4x2 is not permitted to be directly connected to the internet.
+# RFSoC4x2 setup (WAN, direct internet connection)
+This Readme supports setting up your RFSoC4x2, where the RFSoC4x2 is connected directly to the internet--Wide Area Network (WAN). We have an alternative Readme for labs where the RFSoC4x2 is not permitted to be directly connected to the internet.
 
 The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Digital](https://www.realdigital.org/) using AMD’s ZYNQ Ultrascale+ Gen 3 RFSoC ZU48DR chip.  While the ZU48DR has eight digital-to-analog converters (DACs) and analog-to-digital converters (ADCs), the RFSOC4x2 only uses four DACs (5 GSa/s) and two ADCs (9.85 GSa/s). Nonetheless, this number of inputs and outputs is nearly perfect for NV and quantum defect control. However, as the RFSOC4x2 is sold, the ADCs have a high frequency 1GHz high-pass balun inline which is typically too high frequency for our measurements and thus must be modified.
 
@@ -18,19 +18,19 @@ In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Speci
     a. Remove the balun and bypass capacitors<br>
     b. Connect the low frequency differential amplifier<br>
     c. Connect PMOD digital outputs<br>
-    d. Assemble and power on your RFSoC4x2 
+    d. Assemble and power on your RFSoC4x2 <br>
     e. (Optional) Full enclosure
 
-2. Setup lab control computer
-    a. Download/clone QICK-DAWG
-    b. Install necessary packages
+2. Setup lab control computer<br>
+    a. Download/clone QICK-DAWG <br>
+    b. Install necessary packages<br>
 
 3. Upload and install QICK-DAWG and other software to you RFSoC4x2<br>
     a. Flash your microSD card<br>
-    b. Clone QICK-DAWG on your RFSoC4x2 
-    c. Install necessary packages on your RFSOC4x2
-    d. Run the Pyro server to remotely connect to QICK and the RFSoC4x2
-    e. (Supplemental) Establish SSH Remote Connection with the RFSoC4x2
+    b. Clone QICK-DAWG on your RFSoC4x2 <br>
+    c. Install necessary packages on your RFSOC4x2<br>
+    d. Run the Pyro server to remotely connect to QICK and the RFSoC4x2<br>
+    e. (Supplemental) Establish SSH Remote Connection with the RFSoC4x2<br>
 
 
 
@@ -42,11 +42,11 @@ In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Speci
 - Low frequency differential amplifier [Texas Instruments LMH5401EVM](https://www.digikey.com/en/products/detail/texas-instruments/LMH5401EVM/5031896?s=N4IgTCBcDaIDIFkASBWALABgIwFEBqCIAugL5A)
 - 3 x DC output voltage supply (+3.2, +0.7, -1.8V for biasing the differential amplifier)
 - SMA cables
-- Ethernet cord(s) (at least one from RFSoC to router)
+- Ethernet cord
 - Micro SD card reader
 ### Software 
 - [Win32DiskImager](https://sourceforge.net/projects/win32diskimager/) for Windows or Disk Manager on MacOS
-- Dependent packages (follow `Installing Necessary Packages` section--included in batch file setup)
[…]
