### 2024-06-14 e4158e7
fixing any hard coded adc channels
