### 2024-06-24 ac96cc4
fixing hard coded adc channel
