+++ docs/source/installation/Readme_LAN.md
@@ -1,7 +1,9 @@
-# RFSoC4x2 setup (WAN, direct internet connection)
-This Readme supports setting up your RFSoC4x2, where the RFSoC4x2 is connected directly to the internet--Wide Area Network (WAN). We have an alternative Readme for labs where the RFSoC4x2 is not permitted to be directly connected to the internet.
+# RFSoC4x2 Setup (LAN - managed router connection)
+
+This Readme supports setting up your RFSoC4x2 where the RFSoC4x2 is not connected directly to the internet. Instead, here, the RFSoC4x2 is connected to a lab control computer through a managed router connection. There is an alternative Readme in the installation folder for labs where the RFSoC4x2 is permitted to be connected directly to the internet. 
 
 The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Digital](https://www.realdigital.org/) using AMD’s ZYNQ Ultrascale+ Gen 3 RFSoC ZU48DR chip.  While the ZU48DR has eight digital-to-analog converters (DACs) and analog-to-digital converters (ADCs), the RFSOC4x2 only uses four DACs (5 GSa/s) and two ADCs (9.85 GSa/s). Nonetheless, this number of inputs and outputs is nearly perfect for NV and quantum defect control. However, as the RFSOC4x2 is sold, the ADCs have a high frequency 1GHz high-pass balun inline which is typically too high frequency for our measurements and thus must be modified.
+
 
 
 <p align="center">
@@ -14,39 +16,36 @@ The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Di
 
 In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Specifically, we show how to:
 
+
 1. Setup RFSoC4x2 Hardware<br>
     a. Remove the balun and bypass capacitors<br>
-    b. Connect the low frequency differential amplifier<br>
+    b. Connect the low-frequency differential amplifier<br>
     c. Connect PMOD digital outputs<br>
-    d. Assemble and power on your RFSoC4x2 <br>
+    d. Assemble and power on your RFSoC4x2 board
     e. (Optional) Full enclosure
 
-2. Setup lab control computer<br>
-    a. Download/clone QICK-DAWG <br>
-    b. Install necessary packages<br>
-
-3. Upload and install QICK-DAWG and other software to you RFSoC4x2<br>
+2. Upload and install QICK-DAWG and other software to you RFSoC4x2 board<br>
     a. Flash your microSD card<br>
-    b. Clone QICK-DAWG on your RFSoC4x2 <br>
-    c. Install necessary packages on your RFSOC4x2<br>
-    d. Run the Pyro server to remotely connect to QICK and the RFSoC4x2<br>
-    e. (Supplemental) Establish SSH Remote Connection with the RFSoC4x2<br>
+    b. Remotely connect to your RFSoC4x2 through SSH<br>
+    c. Copy required files<br>
+    d. Install necessary packages<br>
 
 
 
-
-# 1. Setup RFSoC4x2 Hardware 
+# 1. Setup RFSoC4x2 hardware 
 ## ***Prerequisites***
 
 - [RFSoC4x2](https://www.xilinx.com/support/university/xup-boards/RFSoC4x2.html) (with 12 volt 50 watt power supply) 
[…]
