+++ .github/workflows/copy_readme_updates_to_rtd_source.yaml
@@ -11,7 +11,7 @@ jobs:
 
     steps:
     - name: Checkout repository
-      uses: actions/checkout@v2
+      uses: actions/checkout@v3
 
     - name: Create destination directory
       run: mkdir -p ../../docs/source/installation
