+++ .github/workflows/copy_readme_updates_to_rtd_source.yaml
@@ -1,30 +1,0 @@
-name: copy readme updates to rtd source 
-
-on: 
-  push:
-    branches: 
-      - docs  # Change to your branch name if different
-
-jobs:
-  copy-file:
-    runs-on: ubuntu-latest
-
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v3
-
-    - name: Create destination directory
-      run: mkdir -p ../../docs/source/installation
-
-    - name: Copy file
-      run: cp ../../../temp.text ../../docs/source/installation/temp.text
-
-    - name: Commit changes
-      run: |
-        git config --global user.name "github-actions[bot]"
-        git config --global user.email "github-actions[bot]@users.noreply.github.com"
-        git add ../../docs/source/installation/temp.text
-        git commit -m "Copy file to destination folder on push"
-        git push
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
