+++ docs/source/installation/Readme_LAN.md
@@ -1,6 +1,6 @@
-# RFSoC4x2 Setup (LAN - managed router connection)
+# RFSoC4x2 Setup (LAN, managed router connection)
 
-This Readme supports setting up your RFSoC4x2 where the RFSoC4x2 is not connected directly to the internet. Instead, here, the RFSoC4x2 is connected to a lab control computer through a managed router connection. There is an alternative Readme in the installation folder for labs where the RFSoC4x2 is permitted to be connected directly to the internet. 
+This Readme supports setting up your RFSoC4x2 where the RFSoC4x2 is not connected directly to the internet. Instead, here, the RFSoC4x2 is connected to a lab control computer through a managed router connection--Local Area Network (LAN). There is an alternative Readme in the installation folder for labs where the RFSoC4x2 is permitted to be connected directly to the internet. 
 
 The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Digital](https://www.realdigital.org/) using AMD’s ZYNQ Ultrascale+ Gen 3 RFSoC ZU48DR chip.  While the ZU48DR has eight digital-to-analog converters (DACs) and analog-to-digital converters (ADCs), the RFSOC4x2 only uses four DACs (5 GSa/s) and two ADCs (9.85 GSa/s). Nonetheless, this number of inputs and outputs is nearly perfect for NV and quantum defect control. However, as the RFSOC4x2 is sold, the ADCs have a high frequency 1GHz high-pass balun inline which is typically too high frequency for our measurements and thus must be modified.
 
@@ -23,12 +23,16 @@ In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Speci
     c. Connect PMOD digital outputs<br>
     d. Assemble and power on your RFSoC4x2 board
     e. (Optional) Full enclosure
+2. Setup lab control computer<br>
+    a. Download/clone QICK-DAWG <br>
+    b. Install necessary packages<br>
 
-2. Upload and install QICK-DAWG and other software to you RFSoC4x2 board<br>
+3. Install QICK-DAWG and other software to you RFSoC4x2 board<br>
     a. Flash your microSD card<br>
     b. Remotely connect to your RFSoC4x2 through SSH<br>
-    c. Copy required files<br>
-    d. Install necessary packages<br>
+    c. Copy required files and install necessary packages<br>
+    d. Connect to the Jupyter server
+    e. Run the Pyro server and remotely connect<br>
 
 
 
@@ -45,7 +49,7 @@ In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Speci
 - Micro SD card reader
 ### Software 
 - [Win32DiskImager](https://sourceforge.net/projects/win32diskimager/) for Windows or Disk Manager on MacOS
-- Dependent packages (follow `Installing Necessary Packages` section--included in batch file setup)
+- Dependent packages (follow `Installing Necessary Packages` section--included the setup batch file)
     - [QICK](https://github.com/openquantumhardware/qick)
     - [Pyro4](https://pypi.org/project/Pyro4/)
     - [Serpent](https://pypi.org/project/serpent/) 
@@ -53,7 +57,8 @@ In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Speci
 
 ## 1a. Remove the balun and bypass capacitors
 
[…]
