### 2024-09-26 a9afccb
Deleted local QICK, updated qick to 2.287 in requirements
