+++ qick/qick_lib/qick/drivers/generator.py
@@ -1,524 +1,0 @@
-"""
-Drivers for signal generators: FPGA blocks that send data to DACs.
-"""
-from pynq.buffer import allocate
-import numpy as np
-from qick import SocIp
-
-class AbsSignalGen(SocIp):
-    """
-    Abstract class which defines methods that are common to different signal generators.
-    """
-    # The DAC channel has a mixer.
-    HAS_MIXER = False
-    # Interpolation factor relating the generator and DAC sampling freqs.
-    FS_INTERPOLATION = 1
-    # Waveform samples per fabric clock.
-    SAMPS_PER_CLK = 1
-    # Maximum waveform amplitude.
-    MAXV = 2**15-2
-    # Scale factor between MAXV and the default maximum amplitude (necessary to avoid overshoot).
-    MAXV_SCALE = 1.0
-
-    # Configure this driver with links to the other drivers, and the signal gen channel number.
-    def configure(self, ch, rf, fs):
-        # Channel number corresponding to entry in the QickConfig list of gens.
-        self.ch = ch
-
-        # RF data converter
-        self.rf = rf
-
-        # DAC sampling frequency.
-        self.cfg['fs'] = fs
-
-        self.cfg['dac'] = self.dac
-
-    def configure_connections(self, soc):
-        self.soc = soc
-
-        # what RFDC port does this generator drive?
-        ((block, port),) = soc.metadata.trace_bus(self.fullpath, 'm_axis')
-        # might need to jump through an axis_register_slice
-        while True:
-            blocktype = soc.metadata.mod2type(block)
-            if blocktype == "usp_rf_data_converter": # we're done
-                break
-            elif blocktype == "axis_register_slice":
-                ((block, port),) = soc.metadata.trace_bus(block, "M_AXIS")
-            elif blocktype == "axis_register_slice_nb":
-                ((block, port),) = soc.metadata.trace_bus(block, "m_axis")
-            else:
-                raise RuntimeError("failed to trace RFDC port for %s - ran into unrecognized IP block %s" % (self.fullpath, block))
-        # port names are of the form 's00_axis'
-        self.dac = port[1:3]
-
-        #print("%s: switch %d, tProc ch %d, DAC tile %s block %s"%(self.fullpath, self.switch_ch, self.tproc_ch, *self.dac))
-
-    def set_nyquist(self, nqz):
-        self.rf.set_nyquist(self.dac, nqz)
-
-    def set_mixer_freq(self, f, ro_ch=None):
-        if not self.HAS_MIXER:
-            raise NotImplementedError("This channel does not have a mixer.")
-        if ro_ch is None:
-            rounded_f = f
-        else:
-            mixercfg = {}
-            mixercfg['f_dds'] = self['fs']
-            mixercfg['b_dds'] = 48
-            fstep = self.soc.calc_fstep(mixercfg, self.soc['readouts'][ro_ch])
-            rounded_f = round(f/fstep)*fstep
-        self.rf.set_mixer_freq(self.dac, rounded_f)
-
-    def get_mixer_freq(self):
-        if not self.HAS_MIXER:
-            raise NotImplementedError("This channel does not have a mixer.")
-        return self.rf.get_mixer_freq(self.dac)
-
-class AbsArbSignalGen(AbsSignalGen):
-    """
-    A signal generator with a memory for envelope waveforms.
-    """
-    # Name of the input driven by the waveform DMA (if applicable).
-    WAVEFORM_PORT = 's0_axis'
-
-    def configure(self, ch, rf, fs):
-        # Define buffer.
[…]
