+++ qick/qick_lib/qick/drivers/tproc.py
@@ -1,860 +1,0 @@
-"""
-Drivers for the QICK timed processor (tProc).
-"""
-from pynq.buffer import allocate
-import numpy as np
-from qick import SocIp
-
-class AxisTProc64x32_x8(SocIp):
-    """
-    AxisTProc64x32_x8 class
-
-    AXIS tProcessor registers:
-    START_SRC_REG
-    * 0 : internal start (using START_REG)
-    * 1 : external start (using "start" input)
-
-    Regardless of the START_SRC, the start logic triggers on a rising edge:
-    A low level arms the trigger (transitions from "end" to "init" state).
-    A high level fires the trigger (starts the program).
-    To stop a running program, see reset().
-
-    START_REG
-    * 0 : init
-    * 1 : start
-
-    MEM_MODE_REG
-    * 0 : AXIS Read (from memory to m0_axis)
-    * 1 : AXIS Write (from s0_axis to memory)
-
-    MEM_START_REG
-    * 0 : Stop.
-    * 1 : Execute operation (AXIS)
-
-    MEM_ADDR_REG : starting memory address for AXIS read/write mode.
-
-    MEM_LEN_REG : number of samples to be transferred in AXIS read/write mode.
-
-    DMEM: The internal data memory is 2^DMEM_N samples, 32 bits each.
-    The memory can be accessed either single read/write from AXI interface. The lower 256 Bytes are reserved for registers.
-    The memory is then accessed in the upper section (beyond 256 bytes). Byte to sample conversion needs to be performed.
-    The other method is to DMA in and out. Here the access is direct, so no conversion is needed.
-    There is an arbiter to ensure data coherency and avoid blocking transactions.
-
-    :param mem: memory address
-    :type mem: int
-    :param axi_dma: axi_dma address
-    :type axi_dma: int
-    """
-    bindto = ['user.org:user:axis_tproc64x32_x8:1.0']
-    REGISTERS = {'start_src_reg': 0,
-                 'start_reg': 1,
-                 'mem_mode_reg': 2,
-                 'mem_start_reg': 3,
-                 'mem_addr_reg': 4,
-                 'mem_len_reg': 5}
-
-    # Number of 32-bit words in the lower address map (reserved for register access)
-    NREG = 64
-
-    def __init__(self, description):
-        """
-        Constructor method
-        """
-        super().__init__(description)
-
-        # Default registers.
-        # start_src_reg = 0   : internal start.
-        # start_reg     = 0   : stopped.
-        # mem_mode_reg  = 0   : axis read.
-        # mem_start_reg = 0   : axis operation stopped.
-        # mem_addr_reg  = 0   : start address = 0.
-        # mem_len_reg   = 100 : default length.
-        self.start_src_reg = 0
-        self.start_reg = 0
-        self.mem_mode_reg = 0
-        self.mem_start_reg = 0
-        self.mem_addr_reg = 0
-        self.mem_len_reg = 100
-
-        # Generics.
-        # data memory address size (log2 of the number of 32-bit words)
-        self.DMEM_N = int(description['parameters']['DMEM_N'])
-        # program memory address size (log2 of the number of 64-bit words, though the actual memory is usually smaller)
-        self.PMEM_N = int(description['parameters']['PMEM_N'])
-
-        self.cfg['dmem_size'] = 2**self.DMEM_N
-
-    # Configure this driver with links to its memory and DMA.
-    def configure(self, mem, axi_dma):
-        # Program memory.
-        self.mem = mem
-
-        # dma
-        self.dma = axi_dma
-
-        self.cfg['pmem_size'] = self.mem.mmio.length//8
-
[…]
