### 2024-09-27 a01aa6e
fix: started to fix NVAverager Program
