+++ qickdawg/nvpulsing/nvaverageprogram.py
@@ -6,7 +6,7 @@ qick.NDAverageProgram
 """
 
 from qick import QickProgram
-from qick.qick_asm import QickRegisterManagerMixin
+from qick.asm_v1 import QickRegisterManagerMixin, AcquireProgram
 from qick.averager_program import AbsQickSweep
 from tqdm.auto import tqdm
 import qickdawg as qd
@@ -21,8 +21,11 @@ import numpy as np
 from typing import List
 from ..util.itemattribute import ItemAttribute
 
+import logging
+logger = logging.getLogger(__name__)
 
-class NVAveragerProgram(QickRegisterManagerMixin, QickProgram):
+
+class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
     """
     NVAveragerProgram class, for experiments that sweep over multiple variables
     in qick-dawg ordered in reps, sweep_n,... sweep_0.
@@ -31,8 +34,6 @@ class NVAveragerProgram(QickRegisterManagerMixin, QickProgram):
 
     Parameters
     --------------------------------------------------------------------------
-    soccfg
-        an instance of QickConfig
     cfg
         an instance of NVConfiguration
 
@@ -50,6 +51,8 @@ class NVAveragerProgram(QickRegisterManagerMixin, QickProgram):
 
     """
 
+    COUNTER_ADDR = 1
+
     def __init__(self, cfg):
         """
         Constructor for the NVAveragerProgram. Make the ND sweep asm commands.
@@ -65,17 +68,20 @@ class NVAveragerProgram(QickRegisterManagerMixin, QickProgram):
             self.rounds = cfg['soft_avgs']
         if "rounds" in cfg:
             self.rounds = cfg['rounds']
+        # reps loop is the outer loop, first-added sweep is innermost loop
+        loop_dims = [cfg['reps'], *self.sweep_axes[::-1]]
+        # average over the reps axis
+        self.setup_acquire(counter_addr=self.COUNTER_ADDR, loop_dims=loop_dims, avg_level=0)
 
     def initialize(self):
         """
-        Abstract method for initializing the program. Should include the instructions that will be executed once at the
-        beginning of the qick program. This is filled in by child classes to make a pulse program.
+        Abstract method for initializing the program and can include any instructions that are executed once at the beginning of the program.
         """
         pass
 
     def body(self):
         """
-        Abstract method for the body of the program. This is filled in by child classes to make a pulse program.
+        Abstract method for the body of the program
         """
         pass
 
@@ -94,7 +100,7 @@ class NVAveragerProgram(QickRegisterManagerMixin, QickProgram):
         """
         Method that makes the assmebly code for an N dimensional sweep pogram. The steps are as follows:
         1. run the overloaded self.initialize() method to initialize the mw and adc channels etc.
-        2. asserts N <7
+        2. asserts N < 5
         3. sets the run count to 0
         4. sets and labels repitition counter
         5. Adds reset and start for each sweep, with label
@@ -111,9 +117,9 @@ class NVAveragerProgram(QickRegisterManagerMixin, QickProgram):
         rep_count = 14  # repetition counter
 
         n_sweeps = len(self.qick_sweeps)
-        if n_sweeps > 7:  # to be safe, only register 15-21 in page 0 can be used as sweep counters
+        if n_sweeps > 5:  # to be safe, only register 15-21 in page 0 can be used as sweep counters
             raise OverflowError(f"too many qick inner loops ({n_sweeps}), run out of counter registers")
-        counter_regs = (np.arange(n_sweeps) + 15).tolist()  # not sure why this has to be a list (np.array doesn't work)
+        counter_regs = (np.arange(n_sweeps) + 17).tolist()  # not sure why this has to be a list (np.array doesn't work)
 
         self.regwi(0, rcount, 0)  # reset total run count
 
[…]
