### 2024-10-09 ebd083b
fix(upgrade qick): upgraded to qick 0.2.289. BREAKING CHANGE
