+++ qickdawg/nvpulsing/nvaverageprogram.py
@@ -11,14 +11,18 @@ from qick.averager_program import AbsQickSweep
 from tqdm.auto import tqdm
 import qickdawg as qd
 
+
 try:
     from rpyc.utils.classic import obtain
 except ModuleNotFoundError:
     def obtain(i):
         return i
-
+    
+import operator
+import functools
 import numpy as np
 from typing import List
+from collections import defaultdict
 from ..util.itemattribute import ItemAttribute
 
 import logging
@@ -120,9 +124,7 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
         if n_sweeps > 5:  # to be safe, only register 15-21 in page 0 can be used as sweep counters
             raise OverflowError(f"too many qick inner loops ({n_sweeps}), run out of counter registers")
         counter_regs = (np.arange(n_sweeps) + 17).tolist()  # not sure why this has to be a list (np.array doesn't work)
-
         self.regwi(0, rcount, 0)  # reset total run count
-
         # set repetition counter and tag
         self.regwi(0, rep_count, self.cfg["reps"] - 1)
         self.label("LOOP_rep")
@@ -157,7 +159,7 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
             sweep_pts.append(swp.get_sweep_pts())
         return sweep_pts
 
-    def acquire(self, soc, load_pulses=True, readouts_per_experiment: int = 1,
+    def acquire(self, load_pulses=True, readouts_per_experiment: int = 1,
                 save_experiments: List = None, start_src: str = "internal",
                 progress=False, remove_offset=True):
         """
@@ -166,8 +168,6 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
 
         Parameters
         ----------
-        soc : QickSoc
-            qick.QickSoc instance
         readouts_per_experiment : int
             int number of readout triggers in the loop body
         load_pulses
@@ -205,22 +205,20 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
         if any([x is None for x in [self.counter_addr, self.loop_dims, self.avg_level]]):
             raise RuntimeError("data dimensions need to be defined with setup_acquire() before calling acquire()")
 
+        # configure tproc for internal/external start
+        qd.soc.start_src(start_src)
+
         n_ro = len(self.ro_chs)
 
-        total_reps = self.expts * self.reps
-        total_count = total_reps * readouts_per_experiment
-        d_buf = np.zeros((n_ro, total_count, 2), dtype=np.int32)
+        total_count = functools.reduce(operator.mul, self.loop_dims)
+        self.d_buf = [np.zeros((*self.loop_dims, nreads, 2), dtype=np.int64) for nreads in self.reads_per_shot]
         self.stats = []
-
-        if 'rounds' not in self.cfg:
-            self.cfg.rounds = 1
 
         # select which tqdm progress bar to show
         hiderounds = True
         hidereps = True
-
         if progress:
-            if self.soft_avgs > 1:
+            if self.cfg.rounds>1:
                 hiderounds = False
             else:
                 hidereps = False
@@ -231,20 +229,22 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
 
         # Actual data acquisition
 
-        for ir in tqdm(range(self.cfg['soft_avgs']), disable=hiderounds):
+        avg_d = None
+        for ir in tqdm(range(self.cfg.rounds), disable=hiderounds):
             # Configure and enable buffer capture.
-            self.config_bufs(soc, enable_avg=True, enable_buf=False)
+            self.config_bufs(qd.soc, enable_avg=True, enable_buf=False)
 
             # Reload data memory.
-            soc.reload_mem()
[…]
