+++ qickdawg/nvpulsing/plintensity.py
@@ -80,7 +80,7 @@ class PLIntensity(NVAveragerProgram):
 
         data = np.mean(data) / self.cfg.readout_integration_treg
 
-        return data
+        return float(data)
 
     def plot_sequence(cfg=None):
       
