### 2024-10-09 efe2b00
fix(init): updated init message
