+++ qickdawg/__init__.py
@@ -2,4 +2,4 @@ from qickdawg.util import *
 from qickdawg.nvpulsing import *
 from qickdawg.fitfunctions import *
 
-# Now using QICK 0.2.160 as of 7/03/2023
+# Now using QICK 0.2.289 as of 10/9/2024
