+++ src/qickdawg/fitfunctions/exponentialdecay.py
@@ -1,0 +1,6 @@
+import numpy as np
+
+
+def exponential_decay(x, a, t, y0):
+
+    return a * np.exp(-x / t) + y0
