### 2024-10-09 74d003c
fix(pyproject): fixed missing quote
