+++ pyproject.toml
@@ -13,7 +13,7 @@ dependencies = [
     "serpent",
     "tqdm",
     "scipy",
-    "qick==0.2.160"
+    "qick==0.2.289"
     ]
 requires-python = ">=3.7"
 description = """
@@ -36,4 +36,4 @@ Repository = "https://github.com/sandialabs/qick-dawg.git"
 exclude = ["qick", "graphics*", "installation*", "jupyter_notebooks*"]
 
 [tool.setuptools.dynamic]
-version = {attr = "qickdawg.__verison__}
+version = {attr = "qickdawg.__verison__"}
