### 2024-10-09 0efa839
install: fixing pyproject.toml
