+++ .github/workflows/python-app.yaml
@@ -39,10 +39,8 @@ jobs:
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Lint with flake8
       run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 .src//qickdawg --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 ./src/qickdawg --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        flake8 ./src/qickdawg --count --max-complexity=10 --max-line-length=127 --statistics
     # - name: Test with pytest
     #   run: |
     #     pytest
