+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+## [0.0.6](https://github.com/sandialabs/qick-dawg/compare/v0.2.0...v0.0.6) (2024-10-09)
+
+
+### Bug Fixes
+
+* **init:** updated init message ([efe2b00](https://github.com/sandialabs/qick-dawg/commit/efe2b0069b64a1edfb3cd989c3ddffb5b6a658fa))
+* **meta:** changed convetional commits and corrected errors for autotesting: ([677d035](https://github.com/sandialabs/qick-dawg/commit/677d0355731042d4142a0ed91cc9e24716fe7648))
+* **pyproject:** fixed missing quote ([74d003c](https://github.com/sandialabs/qick-dawg/commit/74d003cbac645499a1f91a53a4684f74488ca458))
+* started to fix NVAverager Program ([a01aa6e](https://github.com/sandialabs/qick-dawg/commit/a01aa6e1859e35dda76d260663103359143b7715))
+* **upgrade qick:** upgraded to qick 0.2.289. BREAKING CHANGE ([ebd083b](https://github.com/sandialabs/qick-dawg/commit/ebd083bbd998da5d6e7d683302d700ed3c51e693))
+
+
+
