### 2024-10-12 4c4ef57
fix(meta): fixed import so it's no longer broken (#47) [skip ci]
