### 2024-11-19 a73505e
feat(): added simple read out window to calibrate photon ttl levels
