### 2024-11-19 1518307
feat(): added simple read out window to calibrate photon ttl levels
