### 2024-11-19 9e979d9
feat(): updated laser on/off to read out single TTLs from edge counting mode
