+++ src/qickdawg/nvpulsing/laseroff.py
@@ -9,7 +9,7 @@ import numpy as np
 from .nvaverageprogram import NVAveragerProgram
 
 
-def laser_off(config, reps=1, readout_integration_treg=65535):
+def laser_off(config, reps=1, readout_integration_treg=1020):
     '''Sets laser PMOD to low
 
     Parameters
@@ -28,10 +28,13 @@ def laser_off(config, reps=1, readout_integration_treg=65535):
     prog = LaserOff(config)
 
     data = prog.acquire()
-    data = np.mean(data)
-    data /= readout_integration_treg
+    if prog.cfg.edge_counting:
+        return int(data)
+    else:
+        data = np.mean(data)
+        data /= readout_integration_treg
 
-    return float(data)
+        return float(data)
 
 
 class LaserOff(NVAveragerProgram):
@@ -58,10 +61,8 @@ class LaserOff(NVAveragerProgram):
         For LaserOff this simply sets up the adc to integrate for self.cfg.readout_intregration_t#
         """
 
-        self.declare_readout(ch=self.cfg.adc_channel,
-                             freq=0,
-                             length=self.cfg.readout_integration_treg,
-                             sel="input")
+        # Inherited from qd.NVAveragerProgram
+        self.setup_readout()
 
         self.synci(400)  # give processor some time to configure pulses
 
