### 2024-11-19 f63008a
feat(): added NVAverageprogram.setup_readout to setup analog or digital readout
