### 2024-11-19 95c3447
feat(): PLIntensity now works with edge counting, analysis can be rate or total
