### 2024-11-19 d1f25c1
feat(): Lockin-ODMR now has counting mode and analyze digital or analog
