+++ src/qickdawg/util/__init__.py
@@ -1,6 +1,7 @@
 # Functions
 from .cpmgxy8arb import *
 from .compresseddifference import compressed_difference
+from .apply_on_axis_0_n_times import apply_on_axis_0_n_times
 from .absolute import absolute
 from .liveplot import live_plot
 from .intexpscale import int_exp_scale
