+++ src/qickdawg/util/apply_on_axis_0_n_times.py
@@ -1,0 +1,7 @@
+
+
+def apply_on_axis_0_n_times(data, method, n_times):
+
+    for _ in range(n_times):
+        data = method(data, axis=0)
+    return data
