### 2024-11-21 672fc2c
integrating counted rabi
