+++ src/qickdawg/nvpulsing/nvaverageprogram.py
@@ -9,6 +9,7 @@ from qick.asm_v1 import QickRegisterManagerMixin, AcquireProgram
 from qick.averager_program import AbsQickSweep
 from tqdm.auto import tqdm
 import qickdawg as qd
+from ..util import apply_on_axis_0_n_times
 
 
 try:
@@ -443,7 +444,7 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
         self.wait_all()
         self.sync_all(self.cfg.relax_delay_treg + remaining_time)
 
-    def analyze_pulse_sequence_results(self, data):
+    def analyze_analog_pulse_sequence_results(self, data):
         """
         Method that takes in a 1D array of data points from self.acquire() and analyzes the
         results based on the number of reps, rounds, and frequency points
@@ -471,44 +472,107 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
 
         d = ItemAttribute()
 
-        if len(self.data_shape) == 2:
-            d.signal1 = data[:, 0]
-            d.reference1 = data[:, 1]
-            if self.data_shape[-1] == 4:
-                d.signal2 = data[:, 2]
-                d.reference2 = data[:, 3]
-        elif len(self.data_shape) == 3:
-            d.signal1 = data[:, :, 0]
-            d.reference1 = data[:, :, 1]
-            if self.data_shape[-1] == 4:
-                d.signal2 = data[:, :, 2]
-                d.reference2 = data[:, :, 3]
-        elif len(self.data_shape) == 4:
-            d.signal1 = data[:, :, :, 0]
-            d.reference1 = data[:, :, :, 1]
-            if self.data_shape[-1] == 4:
-                d.signal2 = data[:, :, 2]
-                d.reference2 = data[:, :, 3]
+        d.signal1 = data[..., 0]
+        d.reference1 = data[..., 1]
+        if self.data_shape[-1] == 4:
+            d.signal2 = data[..., 2]
+            d.reference2 = data[..., 3]
 
-        if self.data_shape[-1] == 2:
-            d.contrast = ((d.signal1 - d.reference1) / d.reference1 * 100)
-            for _ in range(len(d.contrast.shape) - 1):
-                d.contrast = np.mean(d.contrast, axis=0)
-                d.signal1 = np.mean(d.signal1, axis=0)
-                d.reference1 = np.mean(d.reference1, axis=0)
-        elif self.data_shape[-1] == 4:
-            d.contrast1 = ((d.signal1 - d.reference1) / d.reference1 * 100)
+        n = len(d.signal1.shape) - 1
+
+        if self.cfg.edge_counting is False:
+            ret_type = float
+            func = np.mean
+        else:
+            ret_type = int
+            func = np.sum
+
+        if self.data_shape[-1] >= 2:
+            if self.cfg.edge_counting is False:
+                d.contrast1 = ((d.signal1 - d.reference1) / d.reference1 * 100)
+            else:
+                d.contrast1 = d.signal1 - d.reference1
+            d.contrast1 = apply_on_axis_0_n_times(d.contrast1.astype(float), np.mean, n)
+            d.signal1 = apply_on_axis_0_n_times(d.signal1.astype(float), np.mean, n)
+            d.reference1 = apply_on_axis_0_n_times(d.reference1.astype(float), np.mean, n)
+        
+        if self.data_shape[-1] == 4:
             d.contrast2 = ((d.signal2 - d.reference2) / d.reference2 * 100)
             d.contrast = d.contrast1 - d.contrast2
-            for _ in range(len(d.contrast1.shape) - 1):
-                d.contrast1 = np.mean(d.contrast1, axis=0)
-                d.signal1 = np.mean(d.signal1, axis=0)
 
[…]
