### 2024-12-18 bdb7407
fix(rabi): fixed rabi sequence, dropped 'length' syntax
