### 2024-12-18 776b868
test(readout): ensures that readout does not extend past laser_on time
