### 2024-12-18 b1bdc86
fix(readoutwindow): modified readoutwindow so that it doesn't use sycnall
