### 2025-01-07 e9121ee
feat(client): Added local mode to start client
