### 2025-01-07 9279638
fix(itemattribute): used pypi itemattribute with correct item assignment
