+++ src/qickdawg/nvpulsing/nvconfiguration.py
@@ -6,11 +6,12 @@ setup the pulse sequencing for qick-dawg programs
 """
 
 
-from ..util.itemattribute import ItemAttribute
+from itemattribute import ItemAttribute
 from ..util.intexpscale import int_exp_scale
 
 from math import floor
 import qickdawg as qd
+import numpy as np
 
 
 class NVConfiguration(ItemAttribute):
@@ -38,10 +39,6 @@ class NVConfiguration(ItemAttribute):
 
     Methods
     -------
-    adjust_rounds
-        bandaid method.  An error is typically thrown when more than 1k reps are exectued too quickly
-        this method divides the expereiment into several rounds of less than 1k reps.  Likely slows acuqisition
-
     add_linear_sweep
         method that generates all the attributes required for exectuing a linear sweep
 
@@ -53,12 +50,23 @@ class NVConfiguration(ItemAttribute):
 
         self.soccfg = qd.soccfg
 
+        self.ddr4 = False
+        self.mr = False
+        self.n_ddr4_bins = 10
+
+        self.rounds = 1
+        self.reps = 1
+
+    def __setattribute__(self, name, value):
+
+        super().__setattribute__(self, name, value)
+
     def __setattr__(self, name, value):
         """
         Overloaded class method which handles converting properties
         related to units used by the qick-dawg program
         """
-
+        
         if name.split('_')[-1] == 'tus':
 
             treg = self.soccfg.us2cycles(value)
@@ -104,7 +112,7 @@ class NVConfiguration(ItemAttribute):
 
             self.__dict__[name] = value
             self.__dict__[name.replace('freg', 'fMHz')] = fMHz
-            self.__dict__[name.replace('freg', 'fMHz')] = fMHz / 1000
+            self.__dict__[name.replace('freg', 'fGHz')] = fMHz / 1000
 
         elif name.split('_')[-1] == 'pdegrees':
             preg = self.soccfg.deg2reg(value)
@@ -120,28 +128,6 @@ class NVConfiguration(ItemAttribute):
 
         else:
             self.__dict__[name] = value
-
-    def adjust_rounds(self, reads_per_rep=4):
-        """
-        Take the number of reps (self.reps) and the number of sweep points
-        (self.n_sweep_points) to avoid time outs related to acquiring data too
-        quickly
-
-        Note: this may be fixed with a qick-patch
-        """
-
-        safe_round_points = 1e3
-
-        if 'nsweep_points' not in self:
-            self.nsweep_points = 1
-
-        points_per_round = reads_per_rep * self.nsweep_points * self.reps
-
-        if points_per_round > safe_round_points:
-            reps = safe_round_points // (reads_per_rep * self.nsweep_points)
-
-            self.rounds = int(self.reps // reps)
-            self.reps = int(reps)
 
     def add_linear_sweep(self, name, unit, start, stop, delta=0, nsweep_points=0):
         """
@@ -167,42 +153,93 @@ class NVConfiguration(ItemAttribute):
             (if excluded, must have delta parameter)
         """
 
-        self.__setattr__(name + '_start_' + unit, start)
-        self.__setattr__(name + '_end_' + unit, stop)
-        self.__setattr__(name + '_delta_' + unit, delta)
+        assert np.sum(np.array([delta, nsweep_points]) > 0) == 1, 'Either delta and nsweep_points are required, but not both'
 
-        self.scaling_mode = 'linear'
+        assert unit in [
+            'fMHz', 'fGHz', 'freg',
+            'tus', 'tns', 'treg',
+            'pdeg', 'preg']
+        assert isinstance(nsweep_points, int)
+        if 'reg' in unit:
[…]
