+++ src/qickdawg/nvpulsing/nvaverageprogram.py
@@ -22,7 +22,7 @@ import functools
 import numpy as np
 from typing import List
 from collections import defaultdict
-from ..util.itemattribute import ItemAttribute
+from itemattribute import ItemAttribute
 
 import logging
 logger = logging.getLogger(__name__)
@@ -67,10 +67,12 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
         self.sweep_axes = []
         self.make_program()
         self.reps = cfg['reps']
+
         if "soft_avgs" in cfg:
             self.rounds = cfg['soft_avgs']
         if "rounds" in cfg:
             self.rounds = cfg['rounds']
+        
         # reps loop is the outer loop, first-added sweep is innermost loop
         loop_dims = [cfg['reps'], *self.sweep_axes[::-1]]
         # average over the reps axis
@@ -197,6 +199,11 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
             dimensions for a program with multiple expts/steps: (n_ch, n_reads, n_expts, 2)
         """
 
+        if self.cfg.ddr4 == True:
+            qd.soc.arm_ddr4(ch=self.cfg.ddr4_channel, nt=self.cfg.n_ddr4_chunks)
+        if self.cfg.mr == True:
+            qd.soc.arm_mr(ch=self.cfg.ddr4_channel)
+
         if readouts_per_experiment is not None:
             self.set_reads_per_shot(readouts_per_experiment)
 
@@ -218,7 +225,7 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
         hiderounds = True
         hidereps = True
         if progress:
-            if self.cfg.rounds > 1:
+            if self.rounds > 1:
                 hiderounds = False
             else:
                 hidereps = False
@@ -230,7 +237,7 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
         # Actual data acquisition
 
         # avg_d = None
-        for ir in tqdm(range(self.cfg.rounds), disable=hiderounds):
+        for ir in tqdm(range(self.rounds), disable=hiderounds):
             # Configure and enable buffer capture.
             self.config_bufs(qd.soc, enable_avg=True, enable_buf=False)
 
@@ -262,6 +269,7 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
                         pbar.update(new_points)
 
         return self.d_buf[0][..., 0]
+        # return self.d_buf
 
     def get_data_shape(self, readouts_per_experiment):
         '''
@@ -288,12 +296,15 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
         if self.cfg.reps > 1:
             self.dbuf_shape = [self.cfg.reps] + self.dbuf_shape
 
-        if self.cfg.rounds > 1:
-            self.data_shape = [self.cfg.rounds] + self.dbuf_shape
+        if 'soft_avgs' not in self.cfg:
+            self.cfg.soft_avgs = 1
+
+        if self.cfg.soft_avgs > 1:
+            self.data_shape = [self.cfg.soft_avgs] + self.dbuf_shape
         else:
             self.data_shape = self.dbuf_shape
 
-    def acquire_decimated(self, *arg, **kwarg):
+    def acquire_decimated(self, readouts_per_experiment=None, *arg, **kwarg):
         '''
         Overloaded qick.QickProgram method that drops the Q channel of the time domain readout
 
@@ -308,6 +319,15 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
 
 
         '''
+
+        if self.cfg.ddr4 == True:
[…]
