+++ src/qickdawg/util/__init__.py
@@ -5,5 +5,3 @@ from .absolute import absolute
 from .liveplot import live_plot
 from .intexpscale import int_exp_scale
 from .startclient import start_client
-# Classes
-from .itemattribute import *
