### 2025-01-07 bfb8c05
fix(): updated qick version requirement to include remote fix
