+++ pyproject.toml
@@ -13,7 +13,7 @@ dependencies = [
     "serpent",
     "tqdm",
     "scipy",
-    "qick==0.2.289"
+    "qick==0.2.302"
     ]
 requires-python = ">=3.7"
 description = """
