### 2025-01-07 e0b95ba
test(): added test notebooks for loopback
