### 2025-01-07 a0bd964
fix(graphics): fixed the graphics path error
