+++ src/qickdawg/nvpulsing/rabisweep.py
@@ -196,7 +196,7 @@ class RabiSweep(NVAveragerProgram):
             If None, this plots the squence with configuration labels
             If a `.NVConfiguration` object is supplied, the configuraiton value are added to the plot
         '''
-        graphics_folder = os.path.join(os.path.dirname(__file__), '../../graphics')
+        graphics_folder = os.path.join(os.path.dirname(__file__), 'graphics')
         image_path = os.path.join(graphics_folder, 'RABI.png')
 
         if cfg is None:
