+++ src/qickdawg/util/startclient.py
@@ -37,3 +37,10 @@ def start_client(host_ip_address=None, host_port=8888, server_name="myqick", *ar
         from qick import QickSoc
         qd.soc = QickSoc(*arg, **kwarg)
         qd.soccfg = qd.soc
+
+    qd.max_int_time_tus = qd.soc.cycles2us(2 ** 16 - 1)
+    qd.max_int_time_tns = qd.soc.cycles2us(2 ** 16 - 1) *  1000
+    qd.max_int_time_treg = 2 ** 16 - 1
+
+    qd.min_time_tus = qd.soc.cycles2us(1)
+    qd.min_time_tns = qd.soc.cycles2us(1) * 1000
