+++ src/qickdawg/nvpulsing/t1delaysweep.py
@@ -219,7 +219,7 @@ class T1DelaySweep(NVAveragerProgram):
             If None, this plots the squence with configuration labels
             If a `.NVConfiguration` object is supplied, the configuraiton value are added to the plot
         '''
-        graphics_folder = os.path.join(os.path.dirname(__file__), '../../graphics')
+        graphics_folder = os.path.join(os.path.dirname(__file__), 'graphics')
         image_path = os.path.join(graphics_folder, 'T1.png')
 
         if cfg is None:
