+++ src/qickdawg/nvpulsing/readoutwindow.py
@@ -102,6 +102,14 @@ class ReadoutWindow(NVAveragerProgram):
                              length=self.cfg.readout_length_treg,
                              sel="input")
 
+        self.cfg.adcs = [self.cfg.adc_channel]
+
+        if self.cfg.test:
+            self.declare_readout(ch=self.cfg.mw_readout_channel,
+                                freq=self.cfg.mw_fMHz,
+                                length=self.cfg.readout_length_treg)
+            self.cfg.adcs.append(self.cfg.mw_readout_channel)
+
         # Setup pulse defaults microwave
         if self.cfg.mw_pi2_treg > 0:
 
@@ -116,7 +124,10 @@ class ReadoutWindow(NVAveragerProgram):
 
             self.set_pulse_registers(ch=self.cfg.mw_channel)
 
-        self.synci(400)  # give processor some time to configure pulses
+        self.synci(100)  # give processor some time to configure pulses
+        if (self.cfg.ddr4 == True) or (self.cfg.mr == True):
+            self.trigger(ddr4=self.cfg.ddr4, mr=self.cfg.mr, adc_trig_offset=0)
+        self.synci(100)
 
     def body(self):
         '''
@@ -135,8 +146,8 @@ class ReadoutWindow(NVAveragerProgram):
             )
             t += self.cfg.laser_initialize_treg           
 
-        t += self.cfg.relax_delay_treg
-        self.synci(t)
+            t += self.cfg.relax_delay_treg
+            self.synci(t)
 
         t = 0
         if self.cfg.mw_pi2_treg > 4:
@@ -154,7 +165,7 @@ class ReadoutWindow(NVAveragerProgram):
             t += self.cfg.laser_readout_offset_treg
 
         self.trigger(
-            adcs=[self.cfg.adc_channel],
+            adcs=self.cfg.adcs,
             adc_trig_offset=0,
             pins=[self.cfg.laser_gate_pmod],
             t=t,
