+++ src/qickdawg/nvpulsing/__init__.py
@@ -11,6 +11,7 @@ from .readoutwindow import ReadoutWindow
 from .getreadoutwindow import get_readout_window
 from .check_readout import check_readout
 
+from .integrated_readout_window import IntegratedReadoutWindow
 from .rabisweep import RabiSweep
 from .hahnechodelaysweep import HahnEchoDelaySweep
 from .t1delaysweep import T1DelaySweep
