+++ src/qickdawg/nvpulsing/integrated_readout_window.py
@@ -1,0 +1,243 @@
+from .nvaverageprogram import NVAveragerProgram
+from .nvqicksweep import NVQickSweep
+
+import matplotlib.pyplot as plt
+import matplotlib.image as mpimg
+import os
+import numpy as np
+from ..util.itemattribute import ItemAttribute
+from ..util import apply_on_axis_0_n_times
+
+
+class IntegratedReadoutWindow(NVAveragerProgram):
+    '''
+    An NVAveragerProgram class that generates and executes a sequence used
+    to determine the pi (pi/2) pulse lenghts for your experimetanl configuration
+
+    Parameters
+    -------------------------------------------------------------------
+    soccfg
+        instance of qick.QickConfig class
+    cfg
+        instance of qickdawg.NVConfiguration class with attributes:
+        .adc_channel (required)
+            int channel which is reading data 0 or 1
+
+        .mw_channel (required)
+            qick channel that provides microwave excitation
+            0 or 1 for RFSoC4x2
+            0 to 6 for ZCU111 or ZCU216
+        .mw_nqz (required)
+            nyquist zone for microwave generator (1 or 2)
+        .mw_gain (required)
+            gain of micrwave channel, in register values, from 0 to 2**15-1
+
+        .pre_init (required)
+            boolian value that indicates whether to pre-pulse the laser to initialize
+            the spin state
+
+        .relax_delay_treg (required)
+            int that indicates how long to delay between on/off cycles and reps
+            in register units
+        .readout_length_treg (required)
+            int time for which the adc accumulates data
+            the limit is 1020 points for the FPGA buffer
+        .laser_readout_offset_treg (required)
+
+        .laser_gate_pmod(required)
+            int PMOD channel used to trigger laser source
+            0 to 4
+    returns
+        an instances of LockinODMR class with assembly language compiled
+
+
+    Methods
+    -------
+    initialize
+        method that generates the assembly code that setups the adcs &  mw generators,
+        and performs other one-off setps
+    body
+        method that generates the assembly code that exectues in the middle of each sweep
+        and rep
+    plot_sequence
+        generates a plot labeled with self.cfg attributes or the required inputs
+    time_per_rep
+        returns the approximatetime for one rep to complete
+    total_time
+        returns the approximate total time for the entire program to complete
+
+    '''
+    required_cfg = ["adc_channel",
+                    "readout_integration_treg",
+                    "mw_channel",
+                    "mw_nqz",
+                    "mw_gain",
+                    "mw_freg",
+                    "mw_pi2_treg",
+                    "pre_init",
+                    "laser_gate_pmod",
+                    "laser_on_treg",
+                    "adc_trigger_offset_treg",
+                    "relax_delay_treg",
+                    "mw_readout_delay_treg",
+                    "reps",
+                    "readout_reference_start_treg"
+                    ]
+
+    def initialize(self):
+        '''
+        Method that generates the assembly code that is sets up adcs and sources. 
+        For RabiSweep this:
+        configures the adc to acquire points for self.cfg.readout_integration_t#. 
+        configures the microwave channel 
+        configures the sweep parameters
+        initiailzes the spin state with a laser pulse
+        '''
+        self.check_cfg()
+
+        self.setup_readout()
+
+        # configure pulse defaults and initial parameters for microwave
[…]
