+++ src/qickdawg/nvpulsing/check_readout.py
@@ -19,13 +19,14 @@ def check_readout(config, reps=1, readout_integration_treg=1020):
     '''
 
     config.pre_init = False
-    config.readout_length_treg = readout_integration_treg
+    config.readout_integration_treg = readout_integration_treg
     config.mw_pi2_treg = 0
     config.mw_fMHz = 0
     config.laser_initialize_treg = readout_integration_treg
     config.mw_readout_delay_treg = 0
     config.laser_readout_offset_treg = 0
     config.reps = 1
+
 
     prog = ReadoutWindow(config)
     data = prog.acquire_decimated(progress=False)
