### 2025-01-09 2b2b561
fix(): fixing integration of photon counting with newest version
