### 2025-01-15 ea95c23
fix(install): added itemattribute to the requirements list
