+++ pyproject.toml
@@ -13,7 +13,8 @@ dependencies = [
     "serpent",
     "tqdm",
     "scipy",
-    "qick==0.2.302"
+    "qick==0.2.302",
+    "itemattribute"
     ]
 requires-python = ">=3.7"
 description = """
