### 2025-01-15 7b6875b
cleanup(): using setup_readout method
