+++ src/qickdawg/nvpulsing/ramsey.py
@@ -12,7 +12,7 @@ from .nvqicksweep import NVQickSweep
 
 import matplotlib.pyplot as plt
 import matplotlib.image as mpimg
-import os 
+import os
 
 
 class Ramsey(NVAveragerProgram):
@@ -99,19 +99,7 @@ class Ramsey(NVAveragerProgram):
         '''
         self.check_cfg()
 
-        self.declare_readout(ch=self.cfg.adc_channel,
-                             freq=0,
-                             length=self.cfg.readout_integration_treg,
-                             sel="input")
-
-        self.cfg.adcs = [self.cfg.adc_channel]
-
-        if self.cfg.test:
-            self.declare_readout(ch=self.cfg.mw_readout_channel,
-                                freq=self.cfg.mw_fMHz,
-                                length=self.cfg.readout_integration_treg)
-            self.cfg.adcs.append(self.cfg.mw_readout_channel)
-
+        self.setup_readout()
         # Get registers for mw
 
         self.declare_gen(ch=self.cfg.mw_channel, nqz=self.cfg.mw_nqz)        
@@ -205,7 +193,7 @@ class Ramsey(NVAveragerProgram):
         data = super().acquire(readouts_per_experiment=4, *arg, **kwarg)
 
         if raw_data is False:
-            data = self.analyze_pulse_sequence_results(data)
+            data = self.analyze_pulse_sequence(data)
 
         return data
 
