### 2025-01-15 6cadb08
fix(t1): swapped pulsing order for consistent analysis
