+++ src/qickdawg/nvpulsing/t1delaysweep.py
@@ -176,9 +176,8 @@ class T1DelaySweep(NVAveragerProgram):
         ## First pulse sequence
         ## pi(x) - delay - readout
         # pi2(x)
-        self.pulse(ch=self.cfg.mw_channel)
-        self.pulse(ch=self.cfg.mw_channel)
-        self.sync_all()
+        self.synci(self.cfg.mw_pi2_treg * 2)
+
         # delay
         self.sync(self.delay_register.page, self.delay_register.addr)
         # readout
@@ -190,7 +189,9 @@ class T1DelaySweep(NVAveragerProgram):
         ## Second pulse sequence
         ## Nothing - delay - readout
         # pi(x) - off
-        self.synci(self.cfg.mw_pi2_treg * 2)
+        self.pulse(ch=self.cfg.mw_channel)
+        self.pulse(ch=self.cfg.mw_channel)
+        self.sync_all()
         # delay
         self.sync(self.delay_register.page, self.delay_register.addr) 
         # readout
