### 2025-01-15 99f3689
fix(): changed analysis to consistently have contrast as just the difference
