+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+# [1.1.0](https://github.com/sandialabs/qick-dawg/compare/v1.0.2...v1.1.0) (2025-01-15)
+
+
+### Bug Fixes
+
+* changed analysis to consistently have contrast as just the difference ([99f3689](https://github.com/sandialabs/qick-dawg/commit/99f3689193fd52b401bf3cee09f1952e2b7336fa))
+* **client:** fixed start client for remote use ([da3c227](https://github.com/sandialabs/qick-dawg/commit/da3c2276c963aa9c76bbd9d7c68178ad4ae74255))
+* fixing integration of photon counting with newest version ([2b2b561](https://github.com/sandialabs/qick-dawg/commit/2b2b561c756487008b7e4c56eef0b44696a88c99))
+* **graphics:** fixed the graphics path error ([a0bd964](https://github.com/sandialabs/qick-dawg/commit/a0bd96457cf18ea79692b7e475bb2adfe1770714))
+* **install:** added itemattribute to the requirements list ([ea95c23](https://github.com/sandialabs/qick-dawg/commit/ea95c23445867eadc1b8c0477992633d63e7c3c2))
+* **installation:** simplified package installation, only supporting online installs ([7989c24](https://github.com/sandialabs/qick-dawg/commit/7989c249ffe6b410dbbd6b7fa39e393d4a25b44a))
+* **itemattribute:** used pypi itemattribute with correct item assignment ([9279638](https://github.com/sandialabs/qick-dawg/commit/9279638c9bc8ed6e4b199c472c506f9e55cb58db))
+* **pulsing:** small pulsing changes with integrated ddr4 and testing ([1243e80](https://github.com/sandialabs/qick-dawg/commit/1243e80d6f1d9627afd71e34390caa14125c4dba))
+* **t1:** swapped pulsing order for consistent analysis ([6cadb08](https://github.com/sandialabs/qick-dawg/commit/6cadb082b20df814686d5ce70dc8401c4769297c))
+* updated qick version requirement to include remote fix ([bfb8c05](https://github.com/sandialabs/qick-dawg/commit/bfb8c05bb2530416193ddb9f0598999db0bcc4aa))
+
+
+### Features
+
+* added NVAverageprogram.setup_readout to setup analog or digital readout ([f63008a](https://github.com/sandialabs/qick-dawg/commit/f63008a0d3586949d7d911eefccbc9332204a8c1))
+* added simple read out window to calibrate photon ttl levels ([1518307](https://github.com/sandialabs/qick-dawg/commit/1518307846dac81afcc4606da0b73067339839df))
+* added simple read out window to calibrate photon ttl levels ([a73505e](https://github.com/sandialabs/qick-dawg/commit/a73505e80e438c17fc321d48ae5deafeb4cdf8a2))
+* **client:** Added local mode to start client ([e9121ee](https://github.com/sandialabs/qick-dawg/commit/e9121ee119d2ba81568bd14a6e318cfa29ad14a1))
+* **cpmgxy8:** Added cpmgxy8, delay sweep demonstrated ([b361e42](https://github.com/sandialabs/qick-dawg/commit/b361e42a9190cae5599c9dfe2a913e920fce82f1))
+* Lockin-ODMR now has counting mode and analyze digital or analog ([d1f25c1](https://github.com/sandialabs/qick-dawg/commit/d1f25c1bb5800fba44cf702ecb164f1e4864daf2))
+* PLIntensity now works with edge counting, analysis can be rate or total ([95c3447](https://github.com/sandialabs/qick-dawg/commit/95c34478b1f7bb7e3624fb21d4c7f24681c056ff))
[…]
