### 2025-01-17 8b2b19b
feat(photon_counting): confirmed photon counting works for all pulse sequences
