+++ src/qickdawg/nvpulsing/check_readout.py
@@ -27,7 +27,6 @@ def check_readout(config, reps=1, readout_integration_treg=1020):
     config.laser_readout_offset_treg = 0
     config.reps = 1
 
-
     prog = ReadoutWindow(config)
     data = prog.acquire_decimated(progress=False)
 
