+++ src/qickdawg/nvpulsing/nvaverageprogram.py
@@ -517,6 +517,9 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
             d.contrast1 = d.signal1 - d.reference1
             d.contrast2 = d.signal2 - d.reference2
 
+            d.contrast1_percent = d.signal1 - d.reference1
+            d.contrast2_percent = d.signal2 - d.reference2
+
         d.contrast = d.contrast1 - d.contrast2
         d.contrast_percent = d.contrast1_percent - d.contrast2_percent
 
@@ -525,6 +528,12 @@ class NVAveragerProgram(QickRegisterManagerMixin, AcquireProgram):
                 key = key.format(i)
 
                 d[key] = apply_on_axis_0_n_times(d[key].astype(ret_type), func, n)
+
+        if self.cfg.edge_counting is True:
+            d.contrast1_percent = d.contrast1_percent / d.reference1
+            d.contrast2_percent = d.contrast2_percent / d.reference2
+
+            d.contrast_percent = d.contrast1_percent - d.contrast2_percent
 
         d.contrast = apply_on_axis_0_n_times(d.contrast.astype(ret_type), func, n)
         d.contrast_percent = apply_on_axis_0_n_times(d.contrast_percent.astype(ret_type), func, n)
