+++ src/qickdawg/nvpulsing/readoutwindow.py
@@ -103,8 +103,8 @@ class ReadoutWindow(NVAveragerProgram):
 
         if self.cfg.test:
             self.declare_readout(ch=self.cfg.mw_readout_channel,
-                                freq=self.cfg.mw_fMHz,
-                                length=self.cfg.readout_length_treg)
+                                 freq=self.cfg.mw_fMHz,
+                                 length=self.cfg.readout_length_treg)
             self.cfg.adcs.append(self.cfg.mw_readout_channel)
 
         # Setup pulse defaults microwave
@@ -122,7 +122,7 @@ class ReadoutWindow(NVAveragerProgram):
             self.set_pulse_registers(ch=self.cfg.mw_channel)
 
         self.synci(100)  # give processor some time to configure pulses
-        if (self.cfg.ddr4 == True) or (self.cfg.mr == True):
+        if (self.cfg.ddr4 is True) or (self.cfg.mr is True):
             self.trigger(ddr4=self.cfg.ddr4, mr=self.cfg.mr, adc_trig_offset=0)
         self.synci(100)
 
