+++ src/qickdawg/nvpulsing/cpmgxy8nsweep.py
@@ -2,10 +2,6 @@
 from .nvaverageprogram import NVAveragerProgram
 from .nvqicksweep import NVQickSweep
 
-import matplotlib.pyplot as plt
-import matplotlib.image as mpimg
-import os 
-
 
 class CPMGXY8nSweep(NVAveragerProgram):
     '''
@@ -120,7 +116,7 @@ class CPMGXY8nSweep(NVAveragerProgram):
         self.n_cpmg_register = self.new_gen_reg(
             self.cfg.mw_channel,
             name='ncpmg',
-            init_val = 0)
+            init_val=0)
 
         if self.cfg.scaling_mode == 'exponential':
             self.add_sweep(NVQickSweep(
@@ -143,7 +139,7 @@ class CPMGXY8nSweep(NVAveragerProgram):
             assert 0, 'cfg.scaling_mode must be "linear" or "exponential"'
 
         self.synci(100)  # give processor some time to configure pulses
-        if (self.cfg.ddr4 == True) or (self.cfg.mr == True):
+        if (self.cfg.ddr4 is True) or (self.cfg.mr is True):
             self.trigger(ddr4=self.cfg.ddr4, mr=self.cfg.mr, adc_trig_offset=0)
         self.synci(100)
 
@@ -168,7 +164,6 @@ class CPMGXY8nSweep(NVAveragerProgram):
         9. Loop over reps
         10. Loop over rounds
         '''
-
 
         for i, project_phase in enumerate([0, 180]):
             self.set_pulse_registers(ch=self.cfg.mw_channel, phase=self.deg2reg(0))
