+++ src/qickdawg/nvpulsing/t1delaysweep.py
@@ -107,10 +107,9 @@ class T1DelaySweep(NVAveragerProgram):
 
         if self.cfg.test:
             self.declare_readout(ch=self.cfg.mw_readout_channel,
-                                freq=self.cfg.mw_fMHz,
-                                length=self.cfg.readout_integration_treg)
+                                 freq=self.cfg.mw_fMHz,
+                                 length=self.cfg.readout_integration_treg)
             self.cfg.adcs.append(self.cfg.mw_readout_channel)
-
 
         self.declare_gen(ch=self.cfg.mw_channel, nqz=self.cfg.mw_nqz)
 
@@ -147,7 +146,7 @@ class T1DelaySweep(NVAveragerProgram):
                 self.cfg.nsweep_points))
 
         self.synci(100)  # give processor some time to configure pulses
-        if (self.cfg.ddr4 == True) or (self.cfg.mr == True):
+        if (self.cfg.ddr4 is True) or (self.cfg.mr is True):
             self.trigger(ddr4=self.cfg.ddr4, mr=self.cfg.mr, adc_trig_offset=0)
         self.synci(100)
 
