### 2025-01-17 6c3fabe
Photon counting (#50)
Added photon counting and tested for all pulse sequences except for
CPMGXY8-n
Added DarkCounts program
Added check_readout function [skip ci]
