### 2025-01-17 cd13edf
docs(): cleaned up and updated documentation for installation
