+++ installation/Readme_LAN.md
@@ -1,255 +1,0 @@
-# RFSoC4x2 Setup (LAN, managed router connection)
-
-This Readme supports setting up your RFSoC4x2 where the RFSoC4x2 is not connected directly to the internet. Instead, here, the RFSoC4x2 is connected to a lab control computer through a managed router connection--Local Area Network (LAN). There is an alternative Readme in the installation folder for labs where the RFSoC4x2 is permitted to be connected directly to the internet. 
-
-The RFSoC4x2, as shown in the image below, is a board built and sold by [Real Digital](https://www.realdigital.org/) using AMD’s ZYNQ Ultrascale+ Gen 3 RFSoC ZU48DR chip.  While the ZU48DR has eight digital-to-analog converters (DACs) and analog-to-digital converters (ADCs), the RFSOC4x2 only uses four DACs (5 GSa/s) and two ADCs (9.85 GSa/s). Nonetheless, this number of inputs and outputs is nearly perfect for NV and quantum defect control. However, as the RFSOC4x2 is sold, the ADCs have a high frequency 1GHz high-pass balun inline which is typically too high frequency for our measurements and thus must be modified.
-
-
-
-<p align="center">
-    <img src="graphics/RFSoC4x2_base_image.png"
-        alt="RFSoC4x2 with inputs and outputs labelled"
-        width="600px"/>
-
-</p>
-
-
-In this document we outline the setup for using QICK-DAWG with a RFSoC4x2. Specifically, we show how to:
-
-
-1. Setup RFSoC4x2 Hardware<br>
-    a. Remove the balun and bypass capacitors<br>
-    b. Connect the low-frequency differential amplifier<br>
-    c. Connect PMOD digital outputs<br>
-    d. Assemble and power on your RFSoC4x2 board
-    e. (Optional) Full enclosure
-2. Setup lab control computer<br>
-    a. Download/clone QICK-DAWG <br>
-    b. Install necessary packages<br>
-
-3. Install QICK-DAWG and other software to you RFSoC4x2 board<br>
-    a. Flash your microSD card<br>
-    b. Remotely connect to your RFSoC4x2 through SSH<br>
-    c. Copy required files and install necessary packages<br>
-    d. Connect to the Jupyter server
-    e. Run the Pyro server and remotely connect<br>
-
-
-
-# 1. Setup RFSoC4x2 hardware 
-## ***Prerequisites***
-
-- [RFSoC4x2](https://www.xilinx.com/support/university/xup-boards/RFSoC4x2.html) (with 12 volt 50 watt power supply) 
-- Managed router [(example)](https://www.amazon.com/TP-Link-Integrated-Lightening-Protection-TL-R605/dp/B08QTXNWZ1/ref=asc_df_B08QTXNWZ1/?tag=hyprod-20&linkCode=df0&hvadid=475692076734&hvpos=&hvnetw=g&hvrand=3761702075041011209&hvpone=&hvptwo=&hvqmt=&hvdev=c&hvdvcmdl=&hvlocint=&hvlocphy=1022494&hvtargid=pla-1149738264234&psc=1)
-- Low frequency differential amplifier [Texas Instruments LMH5401EVM](https://www.digikey.com/en/products/detail/texas-instruments/LMH5401EVM/5031896?s=N4IgTCBcDaIDIFkASBWALABgIwFEBqCIAugL5A)
-- 3 x DC output voltage supply (+3.2, +0.7, -1.8V for biasing the differential amplifier)
-- SMA cables
-- Computer with Ethernet port or Ethernet adaptor
-- Ethernet cord(s) (at least one from RFSoC to router)
-- Micro SD card reader
-### Software 
[…]
