### 2025-01-29 f405d28
fix(): fixed installation to remove qd install and paths (#53) [skip ci]
