### 2025-06-25 #6380 @phibeck
Hi @egriendeau thanks for the update. I will let the @openjournals/pe-eics know that you wish to withdraw the paper.

Many thanks to @14shreyasp, @ktahar, @sidihamady for agreeing to review and for your time and support with this submission. Although the paper will be withdrawn, we greatly appreciate your availability and hope to be able to consult you again for future submissions.
